% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absmnorm.R
\name{absmnorm}
\alias{absmnorm}
\title{Absolute Normal Moments}
\usage{
absmnorm(n, sd = 1)
}
\arguments{
\item{n}{vector of integers; order(s) of the moment(s).}

\item{sd}{vector of standard deviations.}
}
\value{
a numeric vector of the absolute moments.
}
\description{
Calculate integer order moments of the absolute value of a centered normal
distribution.
}
\details{
The \eqn{n}th absolute moment of a centered normal variable \eqn{Z} with
standard deviation \eqn{\sigma} is given by
\insertCite{walck1996hand}{msbreg}.

\deqn{E\left[|Z|^{n}\right] = \left\lbrace \begin{matrix} (n - 1)!! \sigma^n & \mathrm{ for } \ n \ even, \\  \\ \sqrt{\frac{2}{\pi}} 2^{k} k! \sigma^{n} & \mathrm{ for \ odd} \ n = 2k + 1. \end{matrix} \right.}

For a normal random variable \eqn{X} with mean \eqn{\mu}
and standard deviation \eqn{\sigma}, the general moments are given
in terms of the moments of the standard normal variable
\eqn{Z = \frac{X - \mu}{\sigma}} as

\deqn{E\left[X^{n}\right] = \sum_{k = 0}^n {k \choose n} \mu^n \sigma^{n - k} E\left[Z^{n - k}\right].}

The standard normal moments \eqn{E\left[Z^{n - k}\right]} are zero for odd \eqn{n}
and can be computed using the function \code{absmnorm} for even \eqn{n}.
}
\examples{
# First six order absolute moments of the standard normal
absmnorm (1:6)

}
\references{
\insertAllCited{}
}
