% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bread_estfun.R
\name{bread.msbm}
\alias{bread.msbm}
\alias{bread.meat}
\alias{estfun.msbm}
\title{Empirical Estimating Functions and Bread Matrix}
\usage{
\method{bread}{msbm}(
  x,
  coefficients = x$coefficients,
  fisher.matrix = TRUE,
  adjust = TRUE,
  ...
)

\method{estfun}{msbm}(x, coefficients = x$coefficients, adjust = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{"msbm"}, a fitted MSB model.}

\item{coefficients}{a numeric vector of all MSB model parameters.
When supplied, \code{coefficients} must have finite values,
and the same length as \code{object$coefficients}.}

\item{fisher.matrix}{logical, should the Fisher (expected) information
matrix be used to obtain the bread matrix? Defaults to \code{TRUE}.
The alternative is the observed information matrix.}

\item{adjust}{logical, should the EEFs (for \code{estfun}) or the
bread matrix (for \code{bread}) be adjusted for any penalty
added to the log-likelihood function during model fitting?
Defaults to \code{TRUE}.}

\item{...}{further arguments passed to or from other methods.
None is currently used.}
}
\value{
For \code{estfun}, the empirical estimating functions, an
\eqn{n \times p} matrix where \eqn{n} is the number of observations
and \eqn{p} is the length of \code{coefficients} (the columns are named
after the \code{$coefficients} component of \code{x}). These are the
scores of the individual contributions to the (penalized) log-likelihood
(derivatives with respect to model parameters).

For \code{bread}, a \eqn{p \times p} matrix where \eqn{p} is the length
of \code{coefficients}: an estimate of the expectation of the the negative
hessian of the (penalized) log-likelihood function (Jacobian of the
estimating functions with respect to model parameters).
If \code{fisher.matrix = FALSE}, the result is the observed hessian while
for \code{fisher.matrix = TRUE}, the expected hessian is returned.
The rows and columns are named after the \code{$coefficients} component of
\code{x}.
}
\description{
Functions to build matrix components of sandwich covariance
estimators: \code{estfun} extracts empirical estimating functions
(EEFs) and \code{bread} extracts an estimator of the scaled expected
information matrix from fitted Multistage Binomial (MSB)
model objects.
}
\details{
These are methods for the package \code{sandwich}'s generics
\link[sandwich]{estfun} and \link[sandwich]{bread}, to handle
objects of class \code{"msbm"}. The methods allow to use the
\link[sandwich]{bread}, \link[sandwich]{meat}, and
\link[sandwich]{sandwich} functions and thus construct any
modified sandwich covariance estimator for objects of class
\code{"msbm"} (see for instance
\insertCite{carroll1998sandwich,skene2010analysis;textual}{msbreg}).

The argument \code{adjust} is by default set to \code{adjust = TRUE}
to ensure that the exact estimating functions are returned, in line
with the default \link[sandwich]{estfun} and \link[sandwich]{bread}
methods. However, \code{adjust = FALSE} is required to obtain the
estimating functions used by default by \link[msbreg]{varcov}.
In general, if likelihood inference is the focus, \code{adjust = FALSE}
should be used since a penalty in the likelihood surface generally
leads to an estimate with a reduced curvature as compared to the
maximum likelihood point (if any it exists). The resulting lost of
efficiency can be masked when using \code{adjust = TRUE}.
}
\examples{
data("test1data")

MSBfit = msbreg (y/Total ~ x1 + offset(off1) | x2,
                 weights = Total,
                 data = test1data)

round(vcov (MSBfit, sandwich = TRUE, adjust = TRUE), 5)

\dontrun{
require (sandwich)
round(sandwich (MSBfit), 5)
}

}
\references{
\insertAllCited{}
}
\seealso{
\link[msbreg]{varcov} for variance-covariance matrix estimation.
}
