% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colStats.R
\name{colStats}
\alias{colStats}
\alias{rowStats}
\title{Row and Column Statistics}
\usage{
colStats (x, stat = median, ...)

rowStats (x, stat = median, ...)
}
\arguments{
\item{x}{an array containing numeric, integer or logical values,
or a numeric data frame.}

\item{stat}{a function, or a character naming a function
that takes a vector input and returns a vector of statistics.
The default is \link{median}.

If the function \code{stat} takes additional inputs, they should
be supplied as named arguments through \code{...}. None of such
arguments can however be named as \code{x} or \code{stat}.}

\item{...}{additional arguments passed to the function \code{stat}.
A common argument is \code{na.rm}, a logical indicating if missing
values (including \code{NaN}) should be omitted from the calculations.}
}
\value{
The output from \link{apply}.

This depends on the presence of \code{simplify} in \code{...},
and the number of dimensions of \code{x}.
See \link{apply} for further details.
}
\description{
Form row and column statistics for numeric matrix (or data frames).
}
\details{
The functions are convenience routines, built as analogous of
\link{colMeans} and \link{rowMeans} for any function appropriate
\code{stat}.

These routines are strictly equivalent to the use of \link{apply}
with \code{FUN = stat} and appropriate margins.
As a result, the argument \code{simplify} of \link{apply} can be
passed through \code{...} to these routines.

Arguments in \code{...} cannot have the same name as any of the
other arguments, or arguments of  \link{apply} such as \code{MARGIN}
or \code{FUN}. Care may also be needed to avoid partial matching
(to for instance \code{stat}, \code{MARGIN} or \code{FUN}).
}
\examples{
##** Compute row and column sums for a matrix:
x <- cbind(x1 = 3, x2 = c(4:1, 2:5))
dimnames(x)[[1]] <- letters[1:8]

##* Scalar statistic
require(msbreg)
rowStats (x, stat = min)
colStats (x, stat = max)
colStats (x, stat = sd)

##* Vector statistic
# Quantiles
rowStats (x, stat = quantile,
          probs = c(0.025, 0.05, 0.1, 0.5, 0.9, 0.95, 0.975))

# Self defined function stat
unistats <- function(x) {
  c(mean = mean(x), min = min(x), med = median(x), max = max(x),
    sd = sd(x), IQR = IQR(x))
}

colStats (x, stat = unistats)

rowStats (x, stat = unistats)

}
\seealso{
\link[msbreg]{colVars} and \link[msbreg]{rowVars}.
}
