% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colVars.R
\name{colVars}
\alias{colVars}
\alias{colSds}
\alias{rowVars}
\alias{rowSds}
\title{Row and Column Statistics}
\usage{
colVars(x, na.rm = FALSE, dims = 1L)

colSds(x, na.rm = FALSE, dims = 1L)

rowVars(x, na.rm = FALSE, dims = 1L)

rowSds(x, na.rm = FALSE, dims = 1L)
}
\arguments{
\item{x}{an array of two or more dimensions, containing numeric,
complex, integer or logical values, or a numeric data frame.}

\item{na.rm}{logical. Should missing values (including \code{NaN})
be omitted from the calculations?}

\item{dims}{integer: Which dimensions are regarded as \emph{rows} or \emph{columns}
to operate over. See \link{colMeans}.}
}
\value{
A numeric or complex array of suitable size,
or a vector if the result is one-dimensional.
The dimnames (or names for a vector result)
are taken from the original array \code{x}.
}
\description{
Form row and column variance and standard deviation
for numeric arrays (or data frames).
}
\details{
These are convenience routines,
built as analogous of \link{colMeans} and \link{rowMeans}
for variance (\link{var}) and standard deviation (\link{sd}).
These functions are equivalent to use of \link{apply} with
\code{FUN = var} or \code{FUN = sd} with appropriate margins,
but are a lot faster. See \link{colMeans} for details
on \code{NaN} and \code{NA} handling.

For a complex input, variance or standard deviation is
computed for the real and the imaginary parts separately
and assembled into a complex result.
}
\examples{
##* Compute row and column variances for a matrix:
x <- cbind(x1 = 3, x2 = c(4:1, 2:5))
rowVars(x)
colVars(x)

##* an array
dim(UCBAdmissions)
rowMeans(UCBAdmissions)
rowVars(UCBAdmissions)

rowMeans(UCBAdmissions, dims = 2)
rowVars(UCBAdmissions, dims = 2)

colVars(UCBAdmissions)
colVars(UCBAdmissions, dims = 2)

##* complex case
x <- cbind(x1 = 3 + 2i, x2 = c(4:1, 2:5) - 5i)
x[3, ] <- NA; x[4, 2] <- NA
rowVars(x)
rowVars(x, na.rm = TRUE)

colVars(x)
colVars(x, na.rm = TRUE)

rowSds(x)
rowSds(x, na.rm = TRUE)

colSds(x)
colSds(x, na.rm = TRUE)

}
\seealso{
\link{colMeans}.
}
