% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_glm.R, R/contour_msbm.R
\name{contour.glm}
\alias{contour.glm}
\alias{contour.msbm}
\alias{contours}
\title{Display Fitted Model Contours}
\usage{
\method{contour}{glm}(
  x,
  xlim = NULL,
  ylim = NULL,
  zlim,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  panel = "panel.contourplot",
  default.prepanel = "prepanel.default.levelplot",
  cuts = 10,
  nlevels = 200,
  labels = TRUE,
  contour = TRUE,
  pretty = TRUE,
  region = TRUE,
  colorkey = list(title = zlab),
  ...
)

\method{contour}{msbm}(
  x,
  xname,
  yname,
  xlim = NULL,
  ylim = NULL,
  zlim,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  panel = "panel.contourplot",
  default.prepanel = "prepanel.default.levelplot",
  cuts = 10,
  nlevels = 200,
  labels = TRUE,
  contour = TRUE,
  pretty = TRUE,
  region = TRUE,
  colorkey = list(title = zlab),
  ...
)
}
\arguments{
\item{x}{an object of class (or inheriting from class) "msbm".}

\item{xlim, ylim, zlim, xlab, ylab, zlab, xname, yname}{labels for the contours.}

\item{panel, default.prepanel}{See \link[lattice]{contourplot}.}

\item{cuts}{number of contour levels desired, i.e. the number of levels the
range of z would be divided into..}

\item{nlevels}{number of function evaluation points per covariate.}

\item{labels, contour, pretty, region}{logicals, see \link[lattice]{contourplot}.}

\item{colorkey}{A logical flag specifying whether a colorkey is to be drawn
alongside the plot, or a list describing the colorkey. See
\link[lattice]{contourplot} for a description of the components of a list
\code{colorkey}.}

\item{...}{Further arguments passed \link[lattice]{contourplot}.}
}
\value{
An object of class \code{"trellis"} (returned invisibly).
See \link[lattice]{contourplot} for a description.
}
\description{
Create color contour plots for fitted model objects.
}
\details{
The function is a \link{contour} method for \code{"msbm"} objects.
It uses \link[lattice]{contourplot} from package \code{lattice}.
}
