% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqrbridge.R
\name{dbridge}
\alias{dbridge}
\alias{pbridge}
\alias{qbridge}
\alias{rbridge}
\alias{bridge}
\title{The Logistic Bridge Distribution}
\source{
\code{[dpq]bridge} are calculated directly from the definitions.

\code{rbridge} uses inversion (see \eqn{F^{-1}(p)} in the section \code{'Details'}).
}
\usage{
dbridge (x, location = 0, scale = 1/sqrt(2), log = FALSE)

pbridge (q, location = 0, scale = 1/sqrt(2), lower.tail = TRUE, log.p = FALSE)

qbridge (p, location = 0, scale = 1/sqrt(2), lower.tail = TRUE, log.p = FALSE)

rbridge (n, location = 0, scale = 1/sqrt(2))
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{location}{vector of means.}

\item{scale}{vector of precision parameters in the range \code{(0, 1]} (the
larger \code{scale} is, the lower the variance).}

\item{log, log.p}{logical; if \code{TRUE}, densities and probabilities \code{p}
are given as \code{log(p)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken
to be the number required (\code{n = length(n)}).}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}.}
}
\value{
\code{dbridge} gives the density, \code{pbridge} gives the distribution
function, \code{qbridge} gives the quantile function, and \code{rbridge}
generates random variates.

The length of the result is determined by \code{n} for \code{rbridge}, and
is the maximum of the lengths of the numerical arguments for the other
functions.

The numerical arguments other than \code{n} are recycled to the length of the result.
Only the first elements of the logical arguments are used.

Negative or null \code{scale} (\eqn{\varphi \leq 0}) is an error and returns
\code{NaN}. The same holds for \code{scale > 1}. For \code{scale = 1}, the
functions consider the limit as \eqn{\varphi} approaches \code{1}, i.e. a
point mass at \eqn{\mu} (\code{location}).
}
\description{
Density, distribution function, quantile function and random generation for
the logistic bridge distribution with mean equal to \code{location} and scale
parameter equal to \code{scale}.
}
\details{
The logistic bridge distribution with \code{scale = }\eqn{\varphi} developed
by \insertCite{wang2003matching;textual}{msbreg} and extended to include a
location parameter \code{location = } \eqn{\mu} has density function

\deqn{f(x) = \frac{1}{2\pi} \frac{\sin (\varphi\pi)}{\cosh (\varphi (x - \mu)) + \cos (\varphi\pi)},}

distribution function

\deqn{F(x) = 1 - \frac{1}{\pi\varphi} \left[\frac{\pi}{2} - \arctan{\left( \frac{\exp(\varphi (x - \mu)) + \cos (\varphi\pi)}{\sin (\varphi\pi)} \right) } \right],}

and quantile function

\deqn{F^{-1}(p) = \frac{1}{\varphi} \log \left[ \frac{\sin (\varphi\pi p)}{\sin (\varphi \pi (1 - p))} \right] + \mu.}

It is a long-tailed distribution (as compared to the Gaussian) with mean
\eqn{\mu} and variance \eqn{\pi^2 \left(\varphi^{-2} - 1\right) / 3} \insertCite{wang2003matching}{msbreg}.

When \code{location} or \code{scale} are not specified they assume the default
values of \eqn{0} and \eqn{1/\sqrt{2}}, respectively, which correspond to
zero mean and the same variance as the unit scale logistic distribution
(\eqn{\pi^2/ 3}).
}
\examples{
## Draw the Bridge density with various scales
curve(dbridge(x, scale = 0.25), -10, 10, col = "black", n = 200,
      lty = 1, ylim = c(0, 0.4), xlab = "x value", ylab = "Density at x")

curve(dbridge(x, scale = 0.5), -10, 10, col = "black", add = TRUE, n = 200,
      lty = 2)

curve(dbridge(x, scale = 0.75), -10, 10, col = "black", add = TRUE, n = 200,
      lty = 3)

legend(5, 0.4, legend = paste0("scale = ", c(0.25, 0.50, 0.75)),
      lty = 1:3)

## Draw the cumulative distribution functions of
# Gaussian, Logistic and Bridge distributions
curve(pnorm, -6, 6, col = "black", ylim = c(0, 1), n = 200,
     xlab = "x value", ylab = "P(X < x)")

curve(plogis(x, scale = sqrt(3)/pi), -6, 6, col = "blue",
      add = TRUE, n = 200)

curve(pbridge(x, scale = pi/sqrt(pi^2 + 3)), -6, 6, col = "red",
      add = TRUE, n = 200)

legend(2, 0.8, legend = c("Gaussian", "Logistic", "Bridge"),
      col = c("black", "blue", "red"), lty = c(1,1,1))


## Random variate generation
# Add density curve to histogram
hist(rbridge(500000, scale = 0.6), freq = FALSE,
     ylim = c(0, 0.25), xlim = c(-10, 10),
     breaks = 200, xlab = "x value", ylab = "Density at x")

curve(dbridge(x, scale = 0.6), -10, 10, n = 200, add = TRUE, col = "blue")

# Variance
var(rbridge(500000, 0, scale = 1/sqrt(2)))
pi^2/3 # True variance
}
\references{
\insertAllCited{}
}
