% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleFactorial.R
\name{DoubleFactorial}
\alias{DoubleFactorial}
\title{Double Factorial}
\source{
\link[msbreg]{DoubleFactorials} and \link[msbreg]{LogDoubleFactorials}
are used for \code{n} values below \code{300} and \code{50000} respectively.
The direct computation uses formulae from
\insertCite{davis1972gamma;textual}{msbreg}.
}
\usage{
DoubleFactorial(n, log = FALSE)
}
\arguments{
\item{n}{Integers vector.}

\item{log}{logical, should the logarithm of the double factorial be returned?}
}
\value{
A numeric vector of the (logarithm of the) double factorials of
\code{n} values.
}
\description{
Calculate the (logarithm of the) double factorial of a number.
}
\details{
The double factorial of an integer \eqn{n} is defined as
\insertCite{gould2012double}{msbreg}

\deqn{n!! = n \cdot (n - 2) \cdot (n - 4) \cdot \cdots \cdot 3 \cdot 1} if \eqn{n} is odd, and

\deqn{n!! = n \cdot (n - 2) \cdot (n - 4) \cdot \cdots \cdot 4 \cdot 2}

if \eqn{n} is even. The double factorial satisfies

\deqn{n!! = n \cdot (n - 2)!!}

with the convention \eqn{(-1)!! = 0!! = 1}.
}
\references{
\insertAllCited{}
}
