% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleFactorial.R
\docType{data}
\name{DoubleFactorials}
\alias{DoubleFactorials}
\alias{LogDoubleFactorials}
\title{Internal Objects}
\format{
An object of class \code{"numeric"}. It is of length 300
for \code{DoubleFactorials} and of length 50 000 for
\code{LogDoubleFactorials}.
}
\source{
\insertCite{Smith2019TreeTools;textual}{msbreg}
}
\usage{
data(DoubleFactorials)

data(LogDoubleFactorials)
}
\description{
Numeric vectors for internal use.
}
\details{
\code{DoubleFactorials} is a vector of pre-calculated double
factorials up to 300!!.

\code{LogDoubleFactorials} has the
logarithms of double factorial values up to 50 000!!.
}
\references{
\insertAllCited{}
}
\seealso{
\link[msbreg]{DoubleFactorial}.
}
\keyword{datasets}
