% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqrtcauchy.R
\name{dtcauchy}
\alias{dtcauchy}
\alias{ptcauchy}
\alias{qtcauchy}
\alias{rtcauchy}
\alias{tCauchy}
\title{The Symmetric Truncated Cauchy Distribution}
\source{
\code{[dpq]tcauchy} are calculated using their relation to the standard
Cauchy distribution (see \link[stats]{Cauchy}).

\code{rtcauchy} uses inversion with \code{qtcauchy}.
}
\usage{
dtcauchy(x, location = 0, scale = 1, log = FALSE, tau = stats::qcauchy(0.9999))

ptcauchy(
  q,
  location = 0,
  scale = 1,
  lower.tail = TRUE,
  log.p = FALSE,
  tau = stats::qcauchy(0.9999)
)

qtcauchy(
  p,
  location = 0,
  scale = 1,
  lower.tail = TRUE,
  log.p = FALSE,
  tau = stats::qcauchy(0.9999)
)

rtcauchy(n, location = 0, scale = 1, tau = stats::qcauchy(0.9999))
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{location}{vector of means.}

\item{scale}{vector of positive dispersion parameters (the larger
\code{scale} is, the larger the variance of the distribution).

Setting \code{scale = } \eqn{\sqrt{\frac{\arctan(\tau)}{\tau - \arctan(\tau)}}} with
\eqn{\tau} \code{= tau} gives a unit variance distribution.

The special value \code{scale = NULL} is allowed. This corresponds to
\code{scale } = \eqn{\pi\sqrt{\frac{1}{3} \frac{\arctan(\tau)}{\tau - \arctan(\tau)}}}
which produces a distribution with the same variance as the unit scale
logistic distribution (\eqn{\pi^2/ 3}).}

\item{log, log.p}{logical; if \code{TRUE}, densities and probabilities \code{p}
are given as \code{log(p)}.}

\item{tau}{positive real, truncation parameter \eqn{\tau} of the distribution.
For \code{location = }\eqn{\mu} and \code{scale = }\eqn{\tau}, only real
values in the range \eqn{[-\sigma \tau + \mu, \sigma\tau + \mu]} have
positive probability densities.

The default value is \code{stats::qcauchy(.9999)} which truncates the
standard Cauchy distribution to the central region with \eqn{99.99\%}
probability mass (the approximate truncation region for the standard
distribution is \eqn{[-3183, 3183]}).}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken
to be the number required (\code{n = length(n)}).}
}
\value{
\code{dtcauchy} gives the density, \code{ptcauchy} gives the
distribution function, \code{qtcauchy} gives the quantile function,
and \code{rtcauchy} generates random variates.

The length of the result is determined by \code{n} for \code{rtcauchy}, and
is the maximum of the lengths of the numerical arguments for the other
functions.

The numerical arguments other than \code{n} are recycled to the length of
the result. Only the first elements of the logical arguments are used.

Negative \code{scale} (\eqn{\sigma < 0}) is an error and returns
\code{NaN}. For \code{scale = 0}, the functions consider the limit
as \eqn{\sigma} approaches \code{0}, i.e. a point mass at
\eqn{\mu} (\code{location}).
}
\description{
Density, distribution function, quantile function and random generation for
the symmetric truncated Cauchy distribution with expectation \code{location}
and scale parameter \code{scale}.
}
\details{
The symmetric truncated Cauchy distribution with location parameter
\code{location = } \eqn{\mu}, scale parameter \code{scale = }\eqn{\sigma}
and truncation parameter \code{tau = } \eqn{\tau > 0}
has density function for
\eqn{x \in \left[-\sigma\tau + \mu, \sigma\tau+\mu \right]}
\insertCite{johnson1995continuous}{msbreg}

\deqn{f(x) = \frac{1}{\pi\sigma P} \left[1 + \frac{(x - \mu)^2}{\sigma^2} \right]^{-1} ,}

where \eqn{P = \frac{2}{\pi} \arctan(\tau)} is the probability
in the region \eqn{\left[-\tau, \tau\right]} under the standard Cauchy
density, and \eqn{f(x) = 0} otherwise. The distribution has expectation
\eqn{\mu} and variance \eqn{\sigma^2\left(\frac{\tau}{\arctan(\tau)} - 1\right)}.

When \code{location} or \code{scale} are not specified they assume the default
values of \eqn{0} and \eqn{1} respectively, which correspond to zero mean and
variance \eqn{\frac{\tau}{\arctan(\tau)} - 1} (see
\insertCite{walck1996hand;nobrackets}{msbreg}, page 30).

The logistic distribution is the basis for the widely use \code{logit} link
function. If using the truncated Cauchy distribution as an alternative
heavy tail link function, it can be useful to set the variance of link
to the variance as the unit scale logistic distribution (\eqn{\pi^2/ 3})
so as to make coefficients estimates comparable to logistic regression
coefficients.
Setting \code{scale = NULL} produces a truncated Cauchy distribution with
variance \eqn{\pi^2/ 3}. Note that the exact distribution depends on the
truncation parameter \code{tau}. For large \code{tau}, the distribution has
long but light tail.
}
\examples{
## Draw the truncated Cauchy density (in -10, 10) with various scales
Maxz <- 10
ylim <- max(dtcauchy(0, scale = 1, tau = Maxz),
            dtcauchy(0, scale = NULL, tau = Maxz))

curve(dtcauchy(x, scale = NULL, tau = Maxz), -Maxz, Maxz, col = "red",
      n = 200, ylim = c(0, ylim), xlab = "x value",
      ylab = "Density at x", lty = 4)

curve(dtcauchy(x, scale = 1, tau = Maxz), -Maxz, Maxz, col = "black",
      add = TRUE, n = 200, lty = 1)

curve(dtcauchy(x, scale = 2, tau = Maxz), -Maxz, Maxz, col = "blue",
      add = TRUE, n = 200, lty = 2)

curve(dtcauchy(x, scale = 5, tau = Maxz), -Maxz, Maxz, col = "brown3",
      add = TRUE, n = 200, lty = 3)

legend(5, 0.32, legend = paste0("scale = ", c('NULL', 1, 2, 5)),
       col = c("red", "black", "blue", "brown3"), lty = c(4, 1:3))

## Draw the cumulative distribution function of the truncated Cauchy
#  compared with Logistic, Logistic Bridge, and Gaussian distributions
scale1 <- sqrt(atan(Maxz) / (Maxz - atan(Maxz))) # scale to get unit variance

curve(ptcauchy(x, scale = scale1), -Maxz, Maxz, col = "red", n = 200)

curve(plogis(x, scale = sqrt(3)/pi), -Maxz, Maxz, col = "brown3",
      add = TRUE, n = 200)

curve(pbridge(x, scale = pi/sqrt(pi^2 + 3)), -Maxz, Maxz, col = "blue",
      add = TRUE, n = 200)

curve(pnorm, -Maxz, Maxz, col = "black", ylim = c(0, 1), add = TRUE, n = 200,
     xlab = "x value", ylab = "P(X < x)")

legend(2, 0.8, legend = c("tCauchy", "Logistic", "Bridge", "Gaussian"),
      col = c("red", "brown3", "blue", "black"), lty = c(1,1,1))

# The truncated Cauchy has heavier tails.

## Random variate generation
# Add density curve to histogram
rtcvalues <- rtcauchy(1e+6, scale = 1, tau = Maxz)
hist(rtcvalues, freq = FALSE,
     ylim = c(0, dtcauchy(0, scale = 1, tau = Maxz)),
     xlim = c(-Maxz, Maxz), breaks = 200,
     xlab = "x value", ylab = "Density at x")

curve(dtcauchy(x, scale = 1, tau = Maxz), -Maxz, Maxz, n = 200,
      add = TRUE, col = "red")

# Variance
var(rtcvalues)
Maxz / atan(Maxz) - 1 # True variance


}
\references{
\insertAllCited{}
}
