% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endometrial.R
\docType{data}
\name{endometrial}
\alias{endometrial}
\title{Endometrial Cancer Data}
\format{
An object of class \code{"data.frame"} with 79 rows and 4 variables:

\describe{
\item{\code{NV}}{neovasculization with coding 0 for absent and 1 for present,}

\item{\code{PI}}{pulsality index of arteria uterina,}

\item{\code{EH}}{endometrium height,}

\item{\code{HG}}{histology grade with coding 0 for low grade and 1 for high grade.}
}
}
\source{
Downloaded from \url{https://users.stat.ufl.edu/~aa/glm/data/}.
}
\usage{
data(endometrial)
}
\description{
Histology grade and risk factors for 79 cases of endometrial cancer.
See \insertCite{heinze2002solution;textual}{msbreg} for a presentation of the study.
}
\references{
\insertAllCited{}
}
\seealso{
See examples in \link[msbreg]{test.separation}.
}
\keyword{datasets}
