% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_matrix_default.R, R/fisher_matrix_glm.R,
%   R/fisher_matrix_msbm.R, R/fisher_matrix_msbm_frame.R
\name{fisher.matrix}
\alias{fisher.matrix}
\alias{fisher.matrix.default}
\alias{fisher.matrix.lm}
\alias{fisher.matrix.glm}
\alias{fisher.matrix.msbm}
\alias{fisher.matrix.msbm.frame}
\title{Fisher Information Matrix}
\usage{
fisher.matrix(object, ...)

\method{fisher.matrix}{lm}(object, sigma = NULL, ...)

\method{fisher.matrix}{glm}(object, sigma = NULL, ...)

\method{fisher.matrix}{msbm}(object, coefficients = object$coefficients, sigma = NULL, ...)

\method{fisher.matrix}{msbm.frame}(object, coefficients, link = "logit", sigma = NULL, ...)
}
\arguments{
\item{object}{an object of a class with a \code{fisher.matrix} method.
Typically, a fit (for instance \code{"lm"} and \code{"glm"} class
objects) or a \code{model.frame} like object.}

\item{...}{further arguments passed to or from other methods.}

\item{sigma}{a numeric scalar (positive real), optional
dispersion parameter in the Generalized Linear Model framework
(e.g. residual standard deviation in Linear Models).

When \code{sigma} is missing or \code{NULL} (the default), it is
inferred from the model frame if the model frame includes
a response, otherwise, \code{sigma = 1} is used.}

\item{coefficients}{a numeric vector of all MSB model parameters.}

\item{link}{optional, a \code{"link"} object (as returned by
\link[msbreg]{link}); defaults to \code{logit()}
(see \link[msbreg]{logit}).}
}
\value{
a \eqn{p \times p} matrix where \eqn{p} is the number of
model parameters, typically the number of columns of the matrix
of contrasts from a model frame.

The output inherits from class \code{"fisher.matrix"} (which
encompasses symmetric non-negative definite matrices) and has
attribute:

\item{\code{sigma}}{ numeric scalar, the used \code{sigma} argument.}

For an \code{object} of class \code{"msbm.frame"} with a non \code{NULL}
response component (\code{attr(object, "response") == 1}), the output
has an additional attribute:

\item{\code{score}}{ numeric matrix with \eqn{p} columns (corresponding
to model parameters) and \eqn{n} rows (corresponding to data points in
the MSB model frame); the \eqn{i}th row of \code{score} gives the
contribution of the \eqn{i}th data point to the model score (first
derivative of the log-likelihood function with respect to all model
parameters); this means that \code{colSums(score)} gives the model
score vector.}

There may be other attributes depending on the method used: see
the appropriate documentation.

There is a simple \code{print} method for \code{"fisher.matrix"} class objects.
}
\description{
Compute the Fisher information matrix of a model (represented by
a fit, or a \code{model.frame} like object) given a parameter vector.
}
\details{
This is a generic function to extract the Fisher information
matrix from a model fit or a model frame (given a vector of
model parameters when required).

The default method can summarily handle any \code{object} of a class
possessing a \code{model.frame} method that returns a \code{"model.frame"}
class output (when applied to the input \code{object}). This includes for
instance \link{formula}, \link{terms}, ... objects, in addition to
proper \link{model.frame} objects. The returned information matrix
corresponds to a Linear Model based on the corresponding
\code{"model.frame"}. The default method also handles
objects of classes \link{lm} and \link{glm}.

For the default method, the \code{...} arguments can include \code{data},
\code{subset}, \code{na.action}, \code{drop.unused.levels}, and \code{xlev};
all passed to \link[stats]{model.frame} when supplied.

Objects of class \code{"msbm"} (as built by \link[msbreg]{msbreg})
have a dedicated \code{fisher.matrix} method  which takes a Multistage
Binomial (MSB) model fit and extracts the Fisher information matrix.
When called on an \code{object} of class \code{"msbm.frame"},
\code{fisher.matrix} takes the additional arguments:

\itemize{
\item \code{coefficients} - \emph{required}, a numeric vector of all MSB model
parameters; and

\item \code{link} - \emph{optional}, a \code{"link"} object (as returned by
\link[msbreg]{link}); defaults to \code{logit()}
(see \link[msbreg]{logit}).
}
}
\examples{
##* Generate a dataset
set.seed(10)
df <- data.frame(y = rbinom(15, 1, 0.5),
                 x1 = rnorm(15),
                 x2 = rpois(15, 5))

##* Fisher information matrix from a model frame (linear model assumed)
library(msbreg)
mf <- model.frame(y ~ x1 + x2, data = df)
fisher.matrix (mf) # dispersion estimated using the response 'y'

fisher.matrix (mf, sigma = 1) # dispersion set to one

mf <- model.frame( ~ x1 + x2, data = df)
fisher.matrix (mf) # sigma = 1 if 'y' is missing (and sigma is not specified)

##* Fisher information matrix for a linear model fit
lm1 <- lm (y ~ x1 + x2, data = df)
fisher.matrix (lm1)

##* Fisher information matrix for a generalized linear model fit
glm1 <- glm (y ~ x1 + x2, family = binomial(), data = df)
fisher.matrix (glm1)

##* Fisher information matrix for a MSB model
data(test1data)

#* Build a MSB model frame
msbframe <- msbm.frame (formula = y/Total ~ x1 + offset(off1) | x2,
                        # minimum probability set to zero
                        alpha.formula = ~ 0,
                        # maximum probability is constant, allowed to be < 1
                        lambda.formula = ~ 1,
                        weights = Total,
                        data = test1data, frames = TRUE)

msbframe

#* Model parameter vector
theta0 <- attr(test1data$y,"theta")
theta0

#* Fisher Information Matrix
FImat <- fisher.matrix (msbframe, coefficients = theta0, link = "logit")

#* Print first two parameters
print(FImat[1:2, 1:2], digits = 5)

#* Strip out parameter names to print them all
drop.attr(FImat, radical = "P")

}
\seealso{
See \link[msbreg]{identifiability} for testing identifiability
of regression models.
}
