% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_to_msbm.R
\name{glm.to.msbm}
\alias{glm.to.msbm}
\title{Binomial \code{glm} to \code{msbm}}
\usage{
glm.to.msbm(
  object,
  refit = FALSE,
  criterion = "ML",
  method = "msbm.fit",
  start = object$coefficients,
  control = object$control,
  y = !is.null(object$y),
  frame = FALSE,
  ...
)
}
\arguments{
\item{object}{an object inheriting from the class \code{"glm"},
and fitted with the \link[stats]{binomial} family (that is,
\code{object$family$family = "binomial"}).}

\item{refit}{logical, should a new fit be performed to replace the
\code{coefficient} component (and related quantities) of the
returned value?}

\item{criterion}{fitting criterion, only used if \code{refit = TRUE}
(see \link[msbreg]{msbm.control} and \link[msbreg]{msbreg}).}

\item{method, start, control, y, frame}{arguments for \link[msbreg]{msbreg},
only used if \code{refit = TRUE}.}

\item{...}{further arguments to pass to or from other methods.

Currently, \code{alpha.formula}, \code{lambda.formula}, \code{me.input}
and \code{stage.me} are passed to \link[msbreg]{msbm.frame} when
\code{refit = TRUE} or \code{method = "model.frame"}.}
}
\value{
an "\link[msbreg]{msbm.frame}" class object if
\code{method = "model.frame"}. Otherwise, an object inheriting
from class \code{"glm2msbm"} which is essentially an \code{"msbm"}
class object, but some components of the returned object miss
some attributes present in regular \code{"msbm"} objects.
}
\description{
Function to convert a binomial \code{"glm"} class object into a
\code{"msbm"} class object.
}
\details{
This function aims to provide a simple routine to allow application
of methods developed for the \code{"msbm"} class to \code{"glm"} class
objects.
}
\examples{
data("infert", package = "datasets")

## Logistic regression fit
GLMfit <- glm (case ~ spontaneous,
               data = infert,
               family = binomial())

summary(GLMfit)

## MSB format
MSBfit <- glm.to.msbm (GLMfit)

summary(MSBfit)

}
