% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HQC.R
\name{HQC}
\alias{HQC}
\alias{HQC.default}
\title{Hannan–Quinn Information Criterion}
\usage{
HQC(object, ...)
}
\arguments{
\item{object}{an object of (or inheriting from) a class with a
\code{HQC} method. The default method handles objects
of a class with a \link[stats]{logLik} or an \link[stats]{AIC} method,
or a list of such objects.}

\item{...}{optionally more fitted model objects of a class with a
\code{HQC} method.}
}
\value{
If just one object is provided, a numeric value with the corresponding HQC.

If multiple objects are provided, a \code{data.frame} with rows corresponding
to the objects and columns representing the number of parameters in the model
(\code{df}) and the HQC.
}
\description{
Generic function calculating the Hannan–Quinn Information Criterion for one or
several fitted model objects for which a log-likelihood value can be obtained.
}
\details{
Similar to the \link[stats]{AIC} function, \code{HQC} calculates a criterion
according to the formula \eqn{-2ll + 2log(log(n)) \times n_{pars}}
where \eqn{ll} is the log-likelihood, \eqn{n} is the number of
observations and \eqn{n_{pars}} is the number of estimated parameters
in the fitted model \insertCite{hannan1979determination}{msbreg}.

This is a generic function (\code{S3}). As for \link[stats]{AIC},
methods should however be defined for
the log-likelihood function \link[stats]{logLik} rather than this function:
the action of the default method is to call \link[stats]{logLik} on all the
supplied objects and assemble the results.

The default method is defined as
\code{AIC(object, ..., k = 2 * log(log(nobs(object))))}.
This requires the number of observations to be known: the method looks first
for a \code{'nobs'} attribute on the return value from the \link[stats]{logLik} method,
then looks for a \code{'nobs'} attribute on the supplied \code{object},
and finally tries the \link[stats]{nobs} generic, and if neither succeed
returns \code{HQC} as \code{NA}.

See \link[stats]{AIC} for some theoretical details and potential problems
related to \link[stats]{logLik} not returning the value at the MLE in
several common cases. This applies to "msbm" class objects obtained
using a fitting criterion different from the simple deviance (see
\link[msbreg]{msbm.control}), or a non-standard fitter (see
\link[msbreg]{msbreg}).
}
\examples{
require(msbreg)
data("test1data")

# Fit univariate and bivariate MSB models
MSBfit1 = msbreg (cbind(y, Total - y) ~ x1,
                  data = test1data)

MSBfit2 = msbreg (cbind(y, Total - y) ~ x1 | x2,
                  data = test1data)

# Get HQC for one model
HQC(MSBfit1)

# Get AIC, HQC and BIC for the two models
cbind(AIC(MSBfit1, MSBfit2),
      HQC = HQC(MSBfit1, MSBfit2)[,2],
      BIC = BIC(MSBfit1, MSBfit2)[,2])

}
\references{
\insertAllCited{}
}
\seealso{
\link[msbreg]{rsquared} for computing generalized (pseudo)
coefficients of determination (\eqn{R^2}) to assess model fits.
}
