% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_test.R
\name{lambda.test}
\alias{lambda.test}
\title{Likelihood Ratio Test}
\usage{
lambda.test(
  object,
  lambda = 1,
  alternative = "less",
  simulate.p.value = FALSE,
  B = 2000
)
}
\arguments{
\item{object}{description}

\item{lambda}{scalar in \code{(0, 1]}.}

\item{alternative}{When \code{lambda = 1}, the function always uses
\code{alternative = 'less'}.}

\item{simulate.p.value}{description}

\item{B}{description}
}
\description{
Likelihood ratio test for \eqn{\lambda = \lambda_0} in multistage
binomial models.
}
