% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LlikDevResid_msbm.R
\name{logLik.msbm}
\alias{logLik.msbm}
\alias{deviance.msbm}
\alias{coef.msbm}
\alias{link.msbm}
\alias{rsquared.msbm}
\title{Assessing Multistage Binomial Model Fits}
\usage{
\method{logLik}{msbm}(object, ...)

\method{deviance}{msbm}(object, ...)

\method{coef}{msbm}(object, ...)

\method{link}{msbm}(object, ...)

\method{rsquared}{msbm}(
  object,
  ...,
  method = "KL",
  cor.method = "pearson",
  adjust.size = FALSE
)
}
\arguments{
\item{object}{an object of class \code{msbm}, typically
the result of a call to \link[msbreg]{msbreg}.}

\item{...}{further arguments passed to or from other methods.}

\item{method}{character indicating the desired pseudo R squared statistic(s).
Any of the following alternatives:

\itemize{
\item \code{'KL'} for Kullback–Leibler divergence based
\eqn{R^2} (default), equivalent to the deviance reduction ratio achieved
by the predictors in a fitted model as compared to a model with no predictor;

\item \code{'Nagelkerke'} for Nagelkerke's \eqn{R^2};

\item \code{'COR'} for the squared linear/monotonic correlation
coefficient (computed by \link[stats]{cor}) between the observed
responses and the corresponding fitted values.
}}

\item{cor.method}{character indicating which correlation coefficient
to consider (only used when \code{method = 'COR'}), this is passed
to \link[stats]{cor} as the \code{method} argument
(\code{method = cor.method}).}

\item{adjust.size}{logical, should the pseudo R-squared statistic be adjusted
for model complexity (number of free parameters)? If \code{TRUE}, the returned
pseudo R-squared statistic is \eqn{1 - \frac{n-1}{n-p} (1 - R^2)} where
\eqn{R^2} is the unadjusted statistic, \eqn{n} is the sample size (number of
observations) and \eqn{p} is the number of free model parameters.}
}
\description{
Methods for assessing class \code{msbm} objects.
}
\details{
The functions \code{coef}, \code{deviance} and \code{logLik} mimick the
corresponding methods for "\link[stats]{glm}" class objects.
The \code{link} method mimicks the \link[stats]{family} function.
The function \code{rsquared} is a method for \link[msbreg]{rsquared}.
}
\seealso{
See \link[msbreg]{summary.msbm} for summarizing
a multistage binomial model fit.
}
