% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_me_family.R
\name{make.me.family}
\alias{make.me.family}
\title{Distribution Functions for Measurement Error Families}
\usage{
make.me.family(me.family)
}
\arguments{
\item{me.family}{character; currently one of \code{"norm"}, \code{"t"},
\code{"logis"}, and \code{"bridge"}, corresponding to the Gaussian, Student t.
Logistic and Bridge distribution families, respectively; or \code{"none"}
(no particular distribution).}
}
\value{
An object of class \code{me-family}, a list with components:
\item{pdf}{probability density function of the family.}
\item{cdf}{cumulative distribution function of the family.}
\item{quf}{quantile (inverse cdf) function of the family.}
\item{name}{a name to be used for the family.}

When \code{me.family = 'none'}, the list elements \code{pdf},
\code{cdf}, and \code{qdf} are \code{NULL}.

Otherwise, \code{pdf}, \code{cdf}, and \code{qdf} are
functions. The arguments of these three functions are
\item{eta}{numeric; vector of quantiles,}
\item{p}{numeric; vector of probabilities,}
\item{sd}{numeric; vector of standard deviations,}
\item{log,log.p}{logical; should the \code{log} of density/probability values
be returned? Defaults to \code{FALSE},}
\item{lower.tail}{logical, should the probability in the lower tail be
returned? Defaults to \code{TRUE}.}

For the Student t family, the three functions have each the
additional argument
\item{df}{numeric; vector of degrees of freedom,
which defaults to \code{2.1}.}
}
\description{
This function is used with \link{link} functions in \link{msbreg}.
Given the name of a family, it returns a probability density function (pdf),
a cumulative distribution function (cdf), and a quantile (inverse cdf)
function (quf).
}
\seealso{
\link[msbreg]{link}, \link[msbreg]{msbreg}.
}
