% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqrlmlogis.R
\name{pmixlogis}
\alias{pmixlogis}
\title{The normal location mixture of logistic distributions}
\usage{
pmixlogis(
  x,
  location = 0,
  scale = 1,
  mix.sd = 0,
  lower.tail = TRUE,
  log.p = FALSE,
  N = NULL
)
}
\arguments{
\item{x}{description}

\item{location}{description}

\item{scale}{description}

\item{mix.sd}{description}

\item{lower.tail}{description}

\item{log.p}{description}

\item{N}{description}
}
\description{
The normal location mixture of logistic distribution is the mixed logistic
distribution obtained from continuously mixing logit probabilities using
a normal mixing distribution.
}
\details{
When \code{mix.sd = 0}, \code{pmixlogis} is the common logistic cumulative
distribution function (cdf) \link{plogis}. For \code{mix.sd > 0}, the function
uses one of two methods: either an adaptive numerical integration with
\link{integrate}, or a truncated series approximation to the cdf. Numerical
integration is used when \code{N = NULL} or \code{N = 0} (the default).
When a numerical value is supplied for \code{N}, this value is used as the
truncation point of the series representation to the cdf.
}
