% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_msbm.R
\name{predict.msbm}
\alias{predict.msbm}
\title{Prediction Method for \code{msbm} Objects}
\usage{
\method{predict}{msbm}(
  object,
  newdata = NULL,
  input.me,
  stage.me,
  type = c("response", "prob", "link", "quantile"),
  at = 0.5,
  se.fit = FALSE,
  fisher.matrix = TRUE,
  na.action = na.pass,
  ...
)
}
\arguments{
\item{object}{fitted model object of class \code{"msbm"}, as returned by
\link[msbreg]{msbreg}.}

\item{newdata}{optional, data frame in which to look for variables
with which to predict. If omitted, the original observations are used.}

\item{input.me, stage.me}{optional named list of one sided
\link[stats]{formula}(s) (\code{input.me}) specifying measurement
errors in some numerical predictors, and a one sided
\link[stats]{formula} (\code{stage.me}) specifying measurement
errors in some stages in the formula of the fitted model
\code{object}. See \link[msbreg]{msbm.frame} for details.

The defaults depend on the presence of argument \code{newdata}.

\itemize{
\item \code{newdata} is missing or \code{NULL}: the default
\code{input.me} and \code{stage.me} are the corresponding components of
\code{object$call} (since the same data used for model fitting is
reconsidered, the default assumes that there are measurement errors,
if any).

\item \code{newdata} is supplied and not \code{NULL}: the defaults
are \code{input.me = list()} and \code{stage.me = NULL} which correspond
to no measurement error in any predictor or stage.
}}

\item{type}{character indicating the desired type of prediction. One of:

\itemize{
\item \code{"response"}: vector of fitted means of binomial responses;

\item \code{"prob"}: vector of fitted success probabilities;

\item \code{"link"}: matrix of linear predictors for model components
(each model stage, and the \eqn{\alpha} and \eqn{\lambda} components);

\item \code{"quantile"}: vector of fitted quantile(s) of the binomial
response distribution;
}

The default is \code{type = "response"}.
The option \code{"prob"} is simply the fitted response divided by the
number of trials (weights). Thus, \code{type = "response"} and
\code{type = "prob"} are equivalent if the response is binary
(all weights are one).}

\item{at}{numeric vector indicating the level(s) at which quantiles
should be predicted (only if \code{type = "quantile"}),
defaulting to the median (\code{at = 0.5)}.}

\item{se.fit}{logical switch indicating if standard errors are desired
when  \code{type = "response"},  \code{type = "prob"},
or \code{type = "link"}.
Note that when \code{type = quantile}, no standard error is available.}

\item{fisher.matrix}{logical, should the Fisher (expected) information
matrix be used to obtain standard errors? Defaults to \code{TRUE}.
Only used when \code{se.fit = TRUE}.}

\item{na.action}{function determining what should be done with missing
values in \code{newdata}. The default is to predict \code{NA}.}

\item{...}{further arguments passed to or from other methods.
None is currently used.}
}
\value{
If \code{se.fit = FALSE}, a vector or matrix of predictions.
For \code{type = 'link'}, this is a matrix with a column per model
component (each model stage, and the \eqn{\alpha} and \eqn{\lambda}
components).

If \code{se.fit = TRUE}, a list with two components:

\itemize{
\item \code{fit}, predictions (as for \code{se.fit = FALSE});
\item \code{se.fit}, estimated standard errors.
}
}
\description{
Extract various types of predictions from Multistage Binomial (MSB)
regression models, on the scale of either the response variable,
the success probability (in \code{[0, 1]}), or the linear predictors
in the MSB model.
}
\details{
Variables are first looked for in \code{newdata} and then searched for
in the usual way (which will include the environment of the formula
used in the fit). A warning will be given if the variables found are
not of the same length as those in \code{newdata} if it was supplied.

If the argument \code{newdata} is omitted, predictions are based
on the data used for the fit. In that case, any measurement error
in the data used for the fit is reconsidered for the predictions
(note however that have no impact on link scale predictions, given
that measurement error distributions are symmetric and have zero mean).
To reuse the data used for model fitting but ignore measurement
error in (some of) the predictors, either set \code{input.me = list()}
and \code{stage.me = NULL}, or leave \code{input.me} and \code{stage.me}
unspecified while passing the dataset used for model fitting as
\code{newdata}.

When \code{se.fit = TRUE}, standard error computation is based on
the delta-method, starting from standard error for model estimated
parameters. Since quantile extraction is not a continuous operation
under the binomial distribution, no standard error is available if
\code{type = "quantile"}, and \code{NA}s are returned as standard
errors in that case.
}
\examples{
data(test1data)
attr(test1data$y, "formula")

##* Two-stage logistic model fit with unknown maximum success probability
# A multiplicative intercept included
msbfit <- msbreg (y/Total ~ x1 + offset(off1) | x2,
                   lambda.formula = ~ 1,
                   weights = Total,
                   data = test1data)

summary (msbfit)

#* Expected binomial responses for a few observations
predict(msbfit, test1data[1:10,])

#* Expected success probability and standard errors
predict(msbfit, test1data[1:10,], type = "prob", se.fit = TRUE)


}
