% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_msbm.R
\name{print.msbm}
\alias{print.msbm}
\alias{print.summary.msbm}
\title{Summarizing a Multistage Binomial Model Fits}
\usage{
\method{print}{msbm}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  me = FALSE,
  frame = FALSE,
  ...
)

\method{print}{summary.msbm}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  show.residuals = FALSE,
  frame = FALSE,
  me = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"msbm"} or \code{"summary.msbm"},
typically, a result of a call to \link[msbreg]{msbreg} or
\link[msbreg]{summary.msbm}.}

\item{digits, symbolic.cor, signif.stars, show.residuals}{Same
as for \link[stats]{summary.glm}:
\itemize{
\item \code{digits} the number of significant digits to use when printing;

\item \code{symbolic.cor} logical, if \code{TRUE}, print the correlations
in a symbolic form (see \link[stats]{symnum}) rather than as numbers;

\item \code{signif.stars} logical, if \code{TRUE}, \emph{significance stars}
are printed for each coefficient;

\item \code{show.residuals} logical, if \code{TRUE} then a summary of the
deviance residuals is printed at the head of the output.;
}}

\item{me}{logical, if \code{TRUE}, the measurement error component
of the model is summarized and printed.}

\item{frame}{logical, if \code{TRUE}, the model frame (component
\code{$frame} of the input) is returned/printed.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The \code{print} methods invisibly returns \code{x}, after printing the main
features of \code{x}.
}
\description{
Print a quick summary of a multistage binomial model fit.
These are \link{methods} for classes \code{"msbm"} and
\code{"summary.msbm"}.
}
\details{
These functions mimic the similar \code{print} methods for
objects of classes "\link[stats]{glm}" and "\link[stats]{summary.glm}".
}
\examples{
# Example 1
set.seed(10)
data("test1data")

MSBfit = msbreg (y/Total ~ x1 + offset(off1) | x2,
                 alpha.formula = ~ 0,
                 lambda.formula = ~ 0,
                 weights = Total,
                 data = test1data)
MSBfit # Print coefficient estimates (descriptive)

summary (MSBfit) # Print a more detailed summary of the fit (inference)
}
