% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linkprobit.R
\name{probit}
\alias{probit}
\alias{link}
\alias{logit}
\alias{cloglog}
\alias{robit}
\alias{cauchit}
\alias{tcauchit}
\alias{sprobit}
\title{Link Objects for Model Fitting}
\usage{
probit(me.family = "conjugate")

sprobit(me.family = "conjugate", lambda = 0.75)

robit(me.family = "conjugate", df = 2.1)

logit(me.family = "conjugate")

cloglog(me.family = "conjugate")

cauchit(me.family = "conjugate")

tcauchit(me.family = "conjugate")

link(object, ...)
}
\arguments{
\item{me.family}{a specification for the measurement error model family.
This can be a name/expression, a literal character string, a length-one
character vector, or an object of class \code{"me-family"} (such as generated
by \link{make.me.family}) provided it is not specified \code{via} one of the
standard names such as \code{probit}, \code{sprobit}, \code{robit},
\code{logit}, \code{cloglog}, \code{cauchit}, and \code{tcauchit}.

Currently, the only available character option is the default
\code{me.family = "conjugate"}. This default picks, if it exists, the natural
(or conjugate) measurement distribution family for a given link function
\code{link}: that is the distribution \code{distr} such that the mixture
of the inverse link function (which is a cumulative distribution function)
by \code{distr} is of the same family as this inverse link function.
This requires the existence of first and second order moments for the
distribution corresponding to the link function, a condition which is
satisfied for all the link functions given here, except for the
\code{cauchit} link (see \strong{Details}).}

\item{object}{the function \code{link} accesses the link of \code{object}
which are stored within objects created by modelling functions
(e.g., \code{msbreg}).}

\item{...}{Additional arguments passed to or from other methods.}

\item{df}{degrees of freedom of the Student t distribution (\link[stats]{TDist}),
\eqn{df >2} required to ensure finite variance \eqn{df/(df - 2)}.
The infinite value \code{df = Inf} is allowed. The variance is then the limit 1.}

\item{lambda}{real skewness parameter of the skew normal distribution
(\link[msbreg]{snorm}). Note that the default \code{lambda = 0.75} corresponds
to a right skewed link with skewness (third standardized moment)
approximately equal to \code{0.8173}.}
}
\value{
An object of class \code{"link"} (which has a concise \code{print}
method). A proper \code{"link"} object is a list with elements:

\item{family}{character: the family name of the response (i.e. \code{"binomial"}).}
\item{link}{character: the link name.}
\item{linkfun}{function: the link.}
\item{linkinv}{function: the inverse of the link function.}
\item{mefamilyname}{character: the measurement error distribution family name.}
\item{mefamily}{a measurement error family, object of class \code{"me-family"}
as returned by \link{make.me.family}.}
\item{sigma}{numeric: the standard deviation of the distribution
corresponding to the link function (\link{NaN} if this is undefined).
The returned \code{sigma} element may have some attributes, e.g.,
for the \code{robit} link, \code{sigma} has attribute \code{df},
i.e. the degree of freedom of the corresponding Student t distribution.}
\item{melinkinv}{function: the inverse link function accounting for
measurement errors.}
\item{mu.eta}{ function: first derivative of the inverse-link function with
respect to the linear predictor.}
\item{mu.eta.eta}{ function: second derivative of the inverse-link function with
respect to the linear predictor.}
}
\description{
Link objects provide a convenient way to specify the details of the links
used by model fitting functions such as \link{msbreg}.
}
\details{
The links given here are for binomial model fittings.
If the model include measurement errors, only the conjugate measurement
distribution family, when it exists, is currently available for each link.
For the \code{probit} link, the conjugate family is the Gaussian family
(see \link[stats]{Normal}). This may be explicitly specified by
\code{probit(me.family = "norm")}.
The \code{sprobit} link which is based on the Sicard skew normal
distribution does not have a known conjugate family. Currently, the
conjugate family is given by the default \code{sprobit(me.family = "conjugate")}.
For the \code{logit} link, the conjugate family is the logistic Bridge
distribution (see \link[msbreg]{dbridge}), the explicit specification
is \code{logit(me.family = "bridge")}.

The \code{robit} link is based on the Student t distribution and thus requires
the specification of a degrees of freedom parameter \code{df} which
determines both the shape (tail behavior) and the variance of the link
\insertCite{liu2004robit,kim2008flexible}{msbreg}.
The degrees of freedom parameter must satisfy \code{df} \eqn{>2} to ensure
a finite link variance \eqn{\sigma^2 = df/(df - 2)} (\eqn{1 < \sigma^2 < \infty}).
So the default \code{df = 2.1} corresponds to a heavy tail distribution with
a variance of \code{21}.
Note that the argument \code{df} is a fixed parameter for model fitting.
The conjugate measurement error family for the \code{robit} link is the
Student t family (this assumes that the link and the measurement errors
are uncorrelated but dependent, as always under the multivariate t
distribution).

The \code{cloglog} link is a skewed link with the log-positive stable
distribution as its conjugate measurement error family
\insertCite{hougaard1986survival}{msbreg}.
The explicit specification here is \code{cloglog(me.family = "lstab")}.

The \code{cauchit} link does not have a conjugate measurement error family
since it has no well defined moment.
The \code{tcauchit} link is based on the symmetric truncated Cauchy
distribution which retains \eqn{99.99\%} probability mass of the standard
Cauchy distribution (the approximate truncation region is
\eqn{[-3183, 3183]}).
This truncated Cauchy distribution has well defined moments.
The \code{tcauchit} link thus has a conjugate measurement error family,
with a distribution function corresponding to a mixture of Cauchy distributions.

The function \code{link} is generic.
The default method extracts the \code{$link} component of \code{object}.
If \code{object} has no \code{$link} component, and is of a class with no
\code{link} method, the default method will output \link{NULL}.
}
\note{
The design was inspired by \code{R}'s \link{family} functions.
}
\references{
\insertAllCited{}
}
\seealso{
\link{make.me.family}.
}
