% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.msbm.R
\name{profile.msbm}
\alias{profile.msbm}
\title{Method for Profiling \code{msbm} Objects}
\usage{
\method{profile}{msbm}(
  fitted,
  which = 1:npar,
  alpha = 0.01,
  maxsteps = 10,
  del = zmax/5,
  trace = FALSE,
  test = c("LRT", "Rao"),
  ...
)
}
\arguments{
\item{fitted}{the original fitted model object.}

\item{which}{the original model parameters which should be profiled.
This can be a numeric or character vector.
By default, all parameters are profiled.}

\item{alpha}{real in \code{(0, 1)}, highest significance level allowed for the profile z-statistics.}

\item{maxsteps}{integer, maximum number of points to be used for profiling each parameter.}

\item{del}{positive real, suggested change on the scale of the profile t-statistics.
Default value chosen to allow profiling at about 10 parameter values.}

\item{trace}{logical, should the progress of profiling be reported?}

\item{test}{character, profile Likelihood Ratio test or Rao Score test.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of classes "profile.msbm" and "profile" with an element for each
parameter being profiled. The elements are data-frames with two variables:

\item{\code{par.vals}}{ a matrix of parameter values for each fitted model;}

\item{\code{tau} or \code{z}}{ the profile t or z-statistics (the name
depends on whether there is an estimated dispersion parameter).}
}
\description{
EXPERIMENTAL FUNCTION.
Investigates the profile log-likelihood function
for a fitted model of class "msbm".
}
\details{
The profile z-statistic is defined either as (case \code{test = "LRT"}) the
square root of change in deviance with an appropriate sign, or
(case \code{test = "Rao"}) as the similarly signed square root of the
Rao Score test statistic.
The latter is defined as the squared gradient of the profile log likelihood
divided by the profile Fisher information, but more conveniently calculated
via the deviance of a Gaussian GLM fitted to the residuals of the profiled model.
}
\author{
Chenangnon Tovissode built the function on the model of
\link[stats]{profile.glm}.
}
