% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilelambda.R
\name{profilelambda}
\alias{profilelambda}
\title{Profile log-likelihood for a MSB model asymptote}
\usage{
profilelambda(
  object,
  lambda = NULL,
  step = 0.01,
  eps = 1e-04,
  neg.criterion = TRUE,
  plotit = FALSE,
  interp = NULL,
  xlab = expression(lambda),
  ylab = "Log-Likelihood",
  ci.level = 0.95,
  boundary = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class \code{"msbm"} typically, a result of a call
to \link[msbreg]{msbreg}.}

\item{lambda}{vector of values of \eqn{\lambda}. The default value is
taken around the estimate of \eqn{\lambda} in \code{object}.}

\item{step}{small positive real specifying the steps of \eqn{\lambda}.
Only used when \code{lambda} is \code{NULL}. Defaults to 0.01.}

\item{eps}{small positive real, tolerance for \eqn{\lambda = 0} and
\eqn{\lambda = 1}. Defaults to 1e-04.}

\item{neg.criterion}{logical, should the negative of the criterion
used to fit \code{object} be considered instead of the log-likelihood?
This is relevant only when the MSB model was fitted using a penalized
likelihood (the default for \link{msbreg}).}

\item{plotit}{logical, should the result be plotted? Defaults to \code{FALSE}.}

\item{interp}{logical, should spline interpolation be used when plotting?
Defaults to \code{TRUE} if plotting with \code{lambda} of length less than 200.}

\item{xlab, ylab}{labels of the axes when \code{plotit = TRUE}.
Defaults to \code{'lambda'} and \code{'log-Likelihood'} respectively.}

\item{ci.level}{level of the confidence interval to be shown on the plot
when \code{plotit = TRUE}.}

\item{boundary}{\code{NULL}, or logical, should the \eqn{\bar{\chi}^2} distribution
having a point mass with probability 0.5 at zero be used for the confidence
interval? The alternative is the common \eqn{\chi^2} distribution.
The default depends on the estimate of \eqn{\lambda} in \code{object} and
the \code{lambda} value with the highest profile log-likelihood value.}

\item{...}{additional parameters to be passed to or from other methods.}
}
\value{
A object of class \code{'profile.lambda'} which is a vector (of the same
length as the unique values in \code{lambda}) with the attributes:

\item{\code{max.index}}{ index of the \eqn{\lambda} value corresponding to
the maximum profiled log-likelihood value (or the negative criterion value)
in \code{lambda};}

\item{\code{pci}}{ profiled confidence interval for \eqn{\lambda};}

\item{\code{lambda}}{ the used vector \code{lambda};}

\item{\code{boundary}}{ the used \code{boundary} value (logical);}

\item{\code{ML}}{ the log-likelihood value corresponding to an
unrestricted estimate of the MSB model in \code{object};}

\item{\code{L1}}{ the log-likelihood value corresponding to \eqn{\lambda = 1}.}

When \code{plotit = TRUE}, the output is invisibly returned (not automatically
printed as usually when an output is not assigned).

The \code{'profile.lambda'} class has a \link[base]{plot} method
for visualizing the calculated log-likelihood profile.
}
\description{
Computes and optionally plots profile log-likelihoods for the
asymptote \eqn{\lambda} of a Multistage Binomial model.
}
\details{
This is a routine to compute the log-likelihood profile for a constant
asymptote \eqn{\lambda} in a MSB model (See \link[msbreg]{msbreg-package}
for an introduction to the MSB model). It is possibly time demanding as
it repeatedly fits the model for different \eqn{\lambda} values in \code{lambda}.

The input argument \code{boundary} should be left as the default \code{NULL},
unless you know what you are doing (i.e. how this affects the returned/plotted
confidence interval). Using the wrong \code{boundary} value will make the
interval too conservative (\code{boundary = FALSE} by error) or too liberal
(\code{boundary = TRUE} by error).

If \code{plotit = TRUE}, the function plots the log-likelihood function profiled
for \eqn{\lambda} and indicates a confidence interval (of level \code{ci.level},
defaults to \code{95\%}) around the \eqn{\lambda} values corresponding to the maximum
observed log-likelihood value. If \code{interp = TRUE}, spline interpolation
is used to give a smoother plot.
}
\examples{
## Two-stage logistic model
data(test1data)
attr(test1data$y, "formula")

# Fit and summarize the model
MSBres <- msbreg (y/Total ~ x1 + offset(off1) | x2,
                   lambda.formula = ~ 1, weights = Total,
                   data = test1data)

summary (MSBres)

# Compute the profile log-likelihood and
# plot with the default 95\% confidence interval
PL <- profilelambda(MSBres, plotit = TRUE,
                    ylab = 'Penalized log-likelihood')

# Show a 90\% confidence interval
plot(PL, ci.level = .9,
     ylab = 'Penalized log-likelihood')

}
