% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbivariate.R
\name{rbivariate}
\alias{rbivariate}
\title{Bivariate Random Variates}
\usage{
rbivariate(
  n,
  mu = c(0, 0),
  sigma = c(1, 1),
  rho = 0,
  family = c("unif", "norm", "lnorm"),
  a,
  b,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length is taken
to be the required number.}

\item{mu}{numeric vector of two means.}

\item{sigma}{numeric vector of two standard deviations.}

\item{rho}{numeric scalar in \code{(-1, 1)}, correlation.
Note that for log-normal variates, the range of feasible
correlation may be limited by the mean values in \code{mu},
especially negative correlations are generally limited to a
threshold way above \code{-1}.}

\item{family}{character, the family of bivariate distributions to sample
from.

Currently, one of \code{'unif'} for the uniform (the default),
\code{'norm'} for the normal (Gaussian), and \code{'lnorm'}
for the log-normal family of distributions.}

\item{a, b}{numeric vectors, minima (\code{a}) and maxima (\code{b}).
Only used when \code{family = 'unif'} and both \code{a} and \code{b}
are specified: indicate the ranges of the values of the two uniform
variables: in that case, \code{mu} and \code{sigma} are ignored.}

\item{verbose}{logical, should correlation adjustment for
log-normal distribution be reported?}

\item{...}{further arguments used by additional families.
Currently none is used.}
}
\description{
Sample from common bivariate distributions with mean vector \code{mu},
standard deviation vector \code{sigma} and correlation \code{rho}.
}
\details{
Bivariate normal variates \eqn{(X_1, X_2)} are simulated by
first sampling \eqn{X_1} from a normal distribution, and then
sampling \eqn{X_2} conditional on \eqn{X_1}. Specifically, \eqn{X_1}
is sampled as:

\code{x1 <- rnorm (n = n, mean = mu[1], sd = sigma[1])}.

Then, the conditional mean of \eqn{X_2} given \eqn{X_1 =}\code{ x1} is
\code{mu_X_2.X_1 = mu[2] + rho * sigma[2] * (x1 - mu[1])/sigma[1]}
and the conditional standard deviation is
\code{sigma_X_2.X_1 = sqrt((1 - rho^2)) * sigma[2])}. Thus,
\eqn{X_2} is sampled as:

\code{x2 <- rnorm (n = n, mean = mu_X_2.X_1, sd = sigma_X_2.X_1}.

Bivariate uniform is simulated via \code{Beta(a,1)} after
\insertCite{demirtas2014generating;textual}{msbreg}.

Bivariate log normal is simulated after
\insertCite{mielke1977covariance;textual}{msbreg}.
}
\examples{
## Bivariate Uniform Variates
# Generate a matrix of observations
BIVU <- rbivariate (100000,
                    mu = c(0.5, 2),
                    sigma = c(2, 2),
                    rho = -0.7,
                    family = "unif")
# Sample mean vector
colMeans (BIVU)

# Sample covariance matrix
cov (BIVU)

# Sample correlation
cor (BIVU)[1,2]

# Scatter plot
plot (BIVU)

## Bivariate Normal Variates
# Generate a matrix of observations
BIVN <- rbivariate (100000,
                    mu = c(0.5, 2),
                    sigma = c(2, 2),
                    rho = -0.7,
                    family = "norm")
# Sample mean vector
colMeans (BIVN)

# Sample covariance matrix
cov (BIVN)

# Sample correlation
cor (BIVN)[1,2]

# Scatter plot
plot (BIVN)

## Bivariate Log normal Variates
# Generate a matrix of observations
BIVL <- rbivariate (100000,
                    mu = c(0.5, 2),
                    sigma = c(2, 2),
                    rho = 0.7,
                    family = "lnorm")
# Sample mean vector
colMeans (BIVL)

# Sample covariance matrix
cov (BIVL)

# Sample correlation
cor (BIVL)[1,2]

# Scatter plot
plot (BIVL)

}
\references{
\insertAllCited{}
}
