% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.msbm}
\alias{residuals.msbm}
\title{Extract Model Residuals}
\usage{
\method{residuals}{msbm}(object, type = c("deviance", "pearson", "response"), ...)
}
\arguments{
\item{object}{an object of class \code{'msbm'}, typically
the result of a call to \link[msbreg]{msbreg}.}

\item{type}{the type of residuals to be returned, the
alternative choices being \code{"deviance"} (default),
\code{"pearson"}, and \code{"response"}.
Can be abbreviated.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Residuals extracted from the fit \code{object}.
}
\description{
Methods for extracting model residuals from objects of class \code{'msbm'}.
}
\details{
This is a \link[stats]{residuals} method for \code{'msbm'} class objects,
mimicking the corresponding methods for "\link[stats]{glm}" class
objects (see \link[stats]{residuals.glm}).
The motivation and definitions for the three types of residuals
(and more) can be found in \insertCite{davison1991residuals;textual}{msbreg}.
}
\references{
\insertAllCited{}
}
\seealso{
See \link[msbreg]{summary.msbm} for summarizing
a multistage binomial model fit.
}
