% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NagelkerkeR2.R, R/Rsquared.R
\name{NagelkerkeR2}
\alias{NagelkerkeR2}
\alias{rsquared}
\alias{rsquared.default}
\alias{rsquared.glm}
\title{Calculate Pseudo \eqn{R^2}}
\usage{
NagelkerkeR2(object, ..., adjust.size = FALSE)

rsquared(object, ...)

\method{rsquared}{glm}(
  object,
  ...,
  method = "KL",
  cor.method = "pearson",
  adjust.size = FALSE
)
}
\arguments{
\item{object}{an object of (or inheriting from) a class with a
\code{rsquared} method. The default method handles objects of
class \code{"glm"}, or a list of such objects.}

\item{...}{optional further arguments to be passed to or from other methods.
Generally, more fitted model objects of a class with a \code{rsquared}
method.}

\item{adjust.size}{logical, should the pseudo R-squared statistic be adjusted
for model complexity (number of free parameters)? If \code{TRUE}, the returned
pseudo R-squared statistic is \eqn{1 - \frac{n-1}{n-p} (1 - R^2)} where
\eqn{R^2} is the unadjusted statistic, \eqn{n} is the sample size (number of
observations) and \eqn{p} is the number of free model parameters.}

\item{method}{character indicating the desired pseudo R squared statistic(s).
Any of the following alternatives:

\itemize{
\item \code{'KL'} for Kullback–Leibler divergence based
\eqn{R^2} (default), equivalent to the deviance reduction ratio achieved by the predictors
in a fitted model as compared to a model with no predictor;

\item \code{'Nagelkerke'} for Nagelkerke's \eqn{R^2};

\item \code{'COR'} for the squared linear/monotonic correlation
coefficient (computed by \link[stats]{cor}) between the observed
responses and the corresponding fitted values.
}}

\item{cor.method}{character indicating which correlation coefficient
to consider (only used when \code{method = 'COR'}), this is passed
to \link[stats]{cor} as the \code{method} argument
(\code{method = cor.method}).}
}
\value{
A \code{'data.frame'} with rows corresponding to the supplied
\code{objects} (including \code{...} if any), and a column of \code{R2} for
each provided method, and two additional columns including \code{N} (number
of observations), \code{df} (number of free model parameters).
As such, \code{NagelkerkeR2} always returns a three-column
\code{data.frame}.
}
\description{
Calculate generalized \eqn{R^2} statistics to assess the goodness of fit
(explanative power) of generalized (non)linear models.
EXTEND TO COVER ALL FIT RESULTS WITH an 'nobs' component, a 'rank' component
(or 'df.null' and 'df.residual' components), and a 'deviance' and a
'deviance.null' components; a 'y' and a 'fitted' component for
correlation based measures.
}
\details{
Arguments after \code{...} must be matched exactly (by their names).

The function \code{rsquared} computes generalized R-squared statistics
such as the deviance reduction ratio or the Nagelkerke's \eqn{R^2}.
This is a generic function for which methods can be added to deal
with specific class objects.

The default method for \code{rsquared} handles objects of classes
\code{'glm'}. It computes by default (\code{method = 'KL'})
the deviance based pseudo-\eqn{R^2} of \insertCite{cameron1996r;textual}{msbreg}.
The alternative \code{method = 'Nagelkerke'} computes
\insertCite{nagelkerke1991note;textual}{msbreg}'s pseudo R-squared.

The function \code{NagelkerkeR2} is an alias to compute Nagelkerke's
R-squared. It extends \link[fmsb]{NagelkerkeR2} of library \code{fmsb} to
handle objects of any class with an \code{rsquared} method (e.g.
\code{'msbm'}) in addition to \code{'glm'} class objects.
}
\note{
The original function \link[fmsb]{NagelkerkeR2} was written by Minato Nakazawa.
}
\examples{
## Logistic regression fit
GLMres <- glm (case ~ spontaneous,
               data = infert,
               family = binomial())

summary(GLMres)

# Kullback–Leibler divergence based R^2 (the default)
rsquared (GLMres)

## MRB model fit with a maximum probability limit
require(msbreg)
MSBres <- msbreg (case ~ spontaneous,
                  lambda.formula = ~ 1,
                  data = infert,
                  start = c(as.numeric(GLMres$coefficients), 4),
                  criterion = 'ML')

summary(MSBres)

# Nagelkerke's R^2 for the two fits
NagelkerkeR2 (GLMres, MSBres)

# Nagelkerke's R^2 and Kullback–Leibler divergence based R^2 for the two fits
rsquared (GLMres, MSBres, method = c('Nagelkerke', 'KL'))

}
\references{
\insertAllCited{}
}
\seealso{
\link[stats]{AIC}, \link[stats]{BIC} and \link[msbreg]{HQC} for
using information criteria to assess model fits.
}
