% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.test.msbm.R
\name{score.test.msbm}
\alias{score.test.msbm}
\title{Score Test for Fitted Model Objects}
\usage{
\method{score.test}{msbm}(
  object,
  ...,
  scope,
  steps = 1000,
  type = c("expected", "observed"),
  penalty = TRUE,
  identity.link.lambda = FALSE,
  side = NULL,
  trace = object$control$trace,
  ncores = NULL
)
}
\arguments{
\item{object}{an object of a class with a \code{score.test} method.
Typically, a fit, for instance \code{"lm"}, \code{"glm"} and \code{"msbm"}
class objects.}

\item{...}{further arguments passed to or from other methods.
This may include one or more alternative models of the same class as \code{object};
and arguments such as \code{which} indicating which parameters to test,
\code{type} for using the expected (Fisher) or observed information matrix,
\code{trace} for verbose, and \code{ncores} for the number of cores for
parallel computing.}

\item{type}{character, type of score statistic to compute. The default
(\code{type = "expected"}) uses the Fisher information matrix.
The alternative (\code{type = "observed"}) uses the observed information
matrix.}

\item{penalty}{logical, should the test account for the penalty when a
penalized fitting criterion was used for object?}

\item{side}{\code{NULL} \code{NA}, \code{-1}, or \code{1} indicating if
one-sided derivatives should be used when numerical derivative are required
(only relevant when \code{type = "observed"} and \code{penalty = TRUE}).}

\item{trace}{integer indicating if details should be printed during
computations.}

\item{ncores}{integer, number of threads for parallel processing.}

\item{which}{character or integer indicating which parameters are targeted.
Defaults to \code{which = NULL}, corresponding to the asymptote parameter
\eqn{\lambda} of the Multistage Binomial (MSB) model.
No alternative is currently available.}

\item{null.values}{numeric vector of null values of target parameters.}
}
\value{
An object of class ‘score.test’ that inherits from the class ‘htest’,
a list with components:

The \code{print} method for \code{"score.test"} class objects calls
\code{scoretable}.
}
\description{
Performs score tests comparing a fitted model to alternatives.
}
\details{
This is a generic function to perform the Rao score test or modified versions
on a model fit object.

The method for an \code{object} of class "\link{lm}" or "\link{glm}" proceeds
as follows.
If a unique model is provided, the default method tests for

When

The function \code{scoretable} generates a summary of one or many
\code{score.test} results.
}
\seealso{
See \link[msbreg]{profile.msbm} for computing log-likelihood profiles
for MSB model fits.
}
