% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractscores.R, R/score.test.R
\name{extractscores}
\alias{extractscores}
\alias{score.test}
\alias{score.test.default}
\alias{score.test.lm}
\alias{score.test.glm}
\title{Score Test for Fitted Model Objects}
\usage{
extractscores(object, which = c(1, 3, 5, 6), ...)

score.test(object, ...)

\method{score.test}{lm}(object, ..., scope, steps = 1000, trace = FALSE, ncores = NULL)

\method{score.test}{glm}(object, ..., scope, steps = 1000, trace = FALSE, ncores = NULL)
}
\arguments{
\item{object}{an object of a class with a \code{score.test} method.
Typically, a fit, for instance \code{"lm"} or \code{"glm"} class objects.}

\item{...}{further arguments passed to or from other methods.
This may include one or more objects specifying alternative models.
These may be objects of class "\link{formula}"s, or (non evaluated)
\link{call}s, or depending on the method for the class of \code{object},
objects of the same class as \code{object}.}

\item{scope}{defines the range of alternative models to be examined.
This is used only when \code{...} does not include any model which nests
\code{object}.
This should be either a single formula, or a list containing components
\code{upper} and \code{lower}, both formulae.
See the details for how to specify the formulae and how they are used.}

\item{steps}{the maximum number of null models to be considered.
The default is 1000 (essentially as many as required).
It is typically used to stop the process early when \code{...} is empty
and \code{scope} is a big model.}

\item{trace}{integer indicating if details should be printed during
computations. Larger values may give more detailed information.}

\item{ncores}{integer, number of threads for parallel processing.}
}
\value{
An object that inherits from the class ‘htests’,
a list with components:

Write 'print.htests' to display a table of test results (seee 'stats:::print.htest')

The \code{print} method for \code{"score.test"} class objects calls
\code{scoretable}.
}
\description{
Performs score tests comparing a fitted (null) model to alternatives.
}
\details{
This is a generic function to perform the Rao score test (or modified versions)
on a model fit.
The score test provides a faster alternative to the Wald (obtained from
a call to \link[stats]{summary.glm}) and the likelihood ratio tests.
Indeed, only the null model need be fitted.
The argument \code{...} allows to precisely indicate the desired alternatives
to the model \code{object}.
If instead, a range of alternatives is desired, the argument \code{scope}
allows to delimit it.

The details given below are for an \code{object} of class "\link{lm}" or "\link{glm}".
Note that all variables in all candidate models should be available from the
null model fit \code{object}, unless \code{scope} contains fitted models.

If alternative models are specified through the argument \code{...},
the function only retains those models which contain \code{object} (the null model)
as a nested model. It then computes the score statistics for the null model
versus each of the alternatives.

If no alternative model is given through the argument \code{...}, or none of the
provided alternatives is larger than \code{object}, the function considers
the argument \code{scope}.
The right-hand-side of the \code{upper} component of \code{scope} is always
considered as an alternative model, provided it is larger than \code{object}.
If \code{scope} is a single formula, it specifies the \code{upper} component,
and the \code{lower} model is \code{object}.
Any model smaller than \code{object} is replaced by \code{object}.
If both the \code{lower} and \code{upper} components indicated by \code{scope}
are identical to \code{object}, no computation is performed and a zero statistic
value is returned, comparing \code{object} with itself.
}
\seealso{
See \link[msbreg]{profile.msbm} for computing log-likelihood profiles
for MSB model fits.

\link{score.test.msbm} for the \code{score.test} method for
fitted Multistage Binomial (MSB) models.
}
