% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_predictors.R
\name{sim.predictors}
\alias{sim.predictors}
\title{Matrix of predictors}
\usage{
sim.predictors(
  n = 100,
  xdistr = "unif",
  xmin = -2,
  xmax = 8,
  cor.structure = NULL
)
}
\arguments{
\item{n}{description}

\item{xdistr}{character vector, indicates the distributions to sample from.}

\item{xmin, xmax}{numeric vectors. These indicate the minimum/maximum of the
uniform distribution to sample for \code{xdistr = 'unif'}. For the other
two distributions, the samples \code{x} are re-scaled to
\code{xmin + (x + 2) * (xmax - xmin) / 10} for \code{xdistr = 'norm'} and
to \code{xmin + x * (xmax - xmin) / 10} for \code{xdistr = 'lnorm'}.}

\item{cor.structure}{character, specifies a correlation structure for
the sampled random vector. Defaults to \code{cor.structure = NULL} for
independent components when the dimension of the sampled random vector is
two or more.}
}
\description{
Simulate a matrix of numerical predictors from uniform, normal or
shifted log-normal variates.
}
