% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_msbm_frame.R
\name{simulate.msbm.frame}
\alias{simulate.msbm.frame}
\alias{simulate.msbm}
\title{Simulate Multistage Binomial Responses}
\usage{
\method{simulate}{msbm.frame}(object, nsim = 1, seed = NULL, theta, link = "logit", ...)

\method{simulate}{msbm}(
  object,
  nsim = 1,
  seed = NULL,
  theta = object$coefficients,
  link = object$link,
  ...
)
}
\arguments{
\item{object}{an object of class \code{"msbm"} or \code{"msbm.frame"},
typically returned by the functions \link[msbreg]{msbreg} or
\link[msbreg]{msbm.frame}.}

\item{nsim, seed, ...}{same as for the default method (\link[stats]{simulate}).}

\item{theta}{numeric vector of model parameters. Note that \code{theta}
is required for an \code{object} of class \code{"msbm.frame"}, and must
be of the same length as the \code{$parnames} component of \code{object}.
For an \code{object} of class \code{"msbm"}, \code{theta} is optional,
with the default given by the \code{$coefficients} component of
\code{object}.}

\item{link}{a character giving the name of the link function to be used.
Alternatively, link can be an object of class \code{"link"},
or a function such that \code{link()} returns an object of class \code{"link"}.
Defaults to the \link[msbreg]{logit} \code{link} for an \code{object} of
class \code{"msbm.frame"}. For an \code{object} of class \code{"msbm"},
the default is the \code{$link} component of \code{object}.}
}
\value{
a numeric vector when \code{nsim = 1},
a numeric \code{"matrix"} class object (\code{nsim} columns)
when \code{nsim > 1}. In either case, the returned value has
the two attributes:
\item{\code{"mu"}}{ the vector of success probabiliies used to generate
individual responses in the returned object;}
\item{\code{"seed"}}{ if argument \code{seed} is \code{NULL},
the attribute is the value of \link{.Random.seed} before the
simulation was started; otherwise it is the value of the argument with a
\code{"kind"} attribute with value \code{as.list(RNGkind())}.}
}
\description{
Simulate one or more vectors of responses from
the distribution corresponding to the response in a
Multistage Binomial (MSB) model represented by a \code{"msbm"}
or a \code{"msbm.frame"} class object and a vector
of model parameters.
}
\details{
These are methods of the generic function \link[stats]{simulate} for
classes \code{"msbm"} and \code{"msbm.frame"}.
The functions provide alternatives to the \link[msbreg]{sim.msbdata} when
a \code{"msbm"} or \code{"msbm.frame"} object is already available.

The method for class \code{"msbm"} simply extracts the model frame
(of class \code{"msbm.frame"}), and if not specified, the vector of
estimated model parameters and the \code{link} of the fit in
\code{object}, and passes them to the method for \code{"msbm.frame"}.
}
\examples{
##** Simulated example: two-stage logistic model
data(test1data)
attr(test1data$y, "formula")

##Two-stage logistic model fit with unknown maximum success probability
# A multiplicative intercept included
MSBres <- msbreg (y/Total ~ x1 + offset(off1) | x2,
                   lambda.formula = ~ 1,
                   weights = Total,
                   data = test1data)

summary (MSBres)

## Simulate 1000 replicates of y from the fitted model
y <- simulate (MSBres, nsim = 1000)

## Compare the original response with a summary from the 1000 simulations
pick <- seq(1, by = 100, to = 1000)
cbind(y = test1data$y[pick],
      fitted = round(MSBres$fitted.values[pick] * test1data$Total[pick], 3),
      sim.mean = rowMeans(y[pick,]),
      sim.min = rowStats(y[pick,], stat = min),
      sim.median = rowStats(y[pick,], stat = median),
      sim.max = rowStats(y[pick,], stat = max))

}
\seealso{
\link[stats]{simulate} for the generic and its default.
See also \link[msbreg]{sim.msbdata} to simulate a MSB model when
\code{object} is not yet available.
}
