% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skew}
\alias{skew}
\alias{skewness}
\title{Coefficient of Skewness}
\usage{
skew(x, na.rm = FALSE, type = c("G1", "b1", "g1", "all"))

skewness(x, na.rm = FALSE, type = c("G1", "b1", "g1", "all"))
}
\arguments{
\item{x}{a numeric \code{vector}, \code{matrix} or \code{data.frame}.}

\item{na.rm}{logical, should missing values be removed?}

\item{type}{character, type of skewness to compute. One of \code{'G1'}
(the adjusted Fisher-Pearson standardized moment coefficient which uses both
unbiased variance and third central moment estimates),
\code{'b1'} (which uses unbiased variance estimate, but biased third central
moment estimate), \code{'g1'} (the traditional Fisher-Pearson
coefficient of skewness, which uses both biased variance and third central
moment estimates), and \code{'all'} (all the three previous types).}
}
\value{
a vector or matrix of sample coefficients of skewness.

If argument \code{x} is a \code{vector}, the output is a vector of three skewness
coefficient estimates if \code{type = 'all'}, otherwise, the output is
a scalar, the specified \code{type} estimate of the skewness coefficient.

If argument \code{x} is a \code{matrix} or a \code{data.frame}, the output
is a matrix where each column is the result from skewness applied to each
column of \code{x} separately.

In all cases, the returned object has attributes \code{'se'} (estimated
asymptotic standard error(s) of the skewness estimates, vector or matrix of
the same size as the output) and \code{'n'} (vector of the size(s) of
samples used for each column of \code{x}).
}
\description{
Estimation of the coefficient of skewness.
}
\details{
The skewness is a measure of the asymmetry of a distribution. For a random
variable \eqn{X}, the skewness \eqn{\gamma_1} (the third standardized moment)
is defined as \insertCite{pearson1895contributions,doane2011measuring}{msbreg}

\deqn{\gamma_1 = \frac{E\left[ (X - \mu)^3 \right]}{E\left[ (X - \mu)^2 \right]^{3/2}}}

where \eqn{\mu = E\left[ X \right]} and all expectations are assumed finite.
The population skewness \eqn{\gamma_1} is zero for a \emph{symmetric} distribution,
negative for a \emph{left-skewed} distribution, and positive for a \emph{right-skewed}
distribution.

The three provided sample coefficients of skewness are the common
estimators, but are generally biased estimates of \eqn{\gamma_1}, unless
\eqn{X} follows a normal distribution.

The function \code{skewness} is a simple alias for \code{skew}.
}
\examples{
## Gaussian samples
rGaussian <- rnorm(500000, sd = 1)

skew (rGaussian, type = 'G1')

## Asymmetric Gaussian samples
rAGaussian <- ranorm(500000, scale1 = 0.5, scale2 = 1.5)

skew (rAGaussian, type = 'all')

skew (cbind(rGaussian, rAGaussian), type = 'all')

}
\references{
\insertAllCited{}
}
