% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_msbm.R
\name{summary.msbm}
\alias{summary.msbm}
\title{Summarizing a Multistage Binomial Model Fits}
\usage{
\method{summary}{msbm}(
  object,
  fisher.matrix = TRUE,
  adjust = FALSE,
  sandwich = FALSE,
  se.values = FALSE,
  IC = NULL,
  correlation = FALSE,
  symbolic.cor = FALSE,
  rsquared.method = "KL",
  cor.method = "pearson",
  frame = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class "msbm",
usually, the result of a call to \link[msbreg]{msbreg}.}

\item{fisher.matrix}{logical, should the Fisher (expected) information
matrix be used to obtain standard errors? Defaults to \code{TRUE}.
The alternative is the observed information matrix.}

\item{adjust}{logical, should the covariance matrix be adjusted for
any penalty added to the log-likelihood function during model fitting?
Defaults to \code{FALSE}.}

\item{sandwich}{logical, should the sandwich variance-covariance
estimator be used? Defaults to \code{FALSE}.}

\item{se.values}{logical, should estimated standard errors of
the \code{mu}, \code{alpha}, \code{lambda}, and
\code{alpha.lambda} components (of the input \code{object}) be
calculated and included in the output? Defaults to \code{FALSE}.}

\item{IC}{logical, should the fit with intercept-correction be considered.}

\item{correlation, symbolic.cor}{Same as for \link[stats]{glm}, i.e.
\itemize{
\item \code{correlation}: a logical, if \code{TRUE}, the correlation matrix
of the estimated parameters is returned and printed;

\item \code{symbolic.cor}: a logical, if \code{TRUE}, print the correlations
in a symbolic form (see \link[stats]{symnum}) rather than as numbers.
}}

\item{rsquared.method}{character indicating which pseudo \eqn{R^2}
statistics to include in the model summary.
Passed as \code{method} to \link[msbreg]{rsquared}.}

\item{cor.method}{character indicating which correlation coefficient
to consider (only used when \code{rsquared.method} includes \code{"COR"}).
Passed as \code{cor.method} to \link[msbreg]{rsquared}.}

\item{frame}{logical, if \code{TRUE}, the model frame (component
\code{$frame} of the input) is returned/printed.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function returns an object of class \code{"summary.msbm"},
a list with components:

\item{\code{call}, \code{link}, \code{deviance}, \code{aic},
\code{bic}, \code{hqc}, \code{df.residual}, \code{null.deviance},
\code{df.null}, \code{iter}, \code{converged}, \code{boundary},
\code{na.action}, \code{deviance.resid}, \code{y.resid},
\code{dispersion}, \code{dims}, \code{me}, \code{me.offset},
\code{me.sd} and \code{frame}}{ the corresponding components from
\code{object};}

\item{\code{coefficients}}{ the matrix of coefficients,
standard errors, \code{z}-values and \code{p}-values
(aliased coefficients are omitted in the returned output,
but printed);}

\item{\code{aliased}}{ named logical vector showing if the original coefficients are aliased;}

\item{\code{df}}{ a 3-vector of the rank of the model and the number of
residual degrees of freedom, plus number of coefficients (including aliased ones);}

\item{\code{cov.scaled}}{ the (scaled) estimate of the covariance matrix of
the estimated coefficients;}

\item{\code{correlation}}{ if requested (\code{correlation = TRUE}),
the estimate of the matrix of correlations between the estimated
coefficients;}

\item{\code{symbolic.cor}}{ if \code{correlation = TRUE},
the value of the argument \code{symbolic.cor};}

\item{\code{mu}}{ numeric, the \code{$fitted.values} component of
\code{object}, that is, the fitted individual success probabilities
\eqn{\mu} of the binary responses;}

\item{\code{alpha}, \code{lambda}}{ numeric, the \eqn{\alpha} and
\eqn{\lambda} components (fitted or known values/offsets) of the model:
the vector of individual \eqn{\alpha} or \eqn{\lambda} values is
reduced to a scalar if constant across all individuals;}

\item{\code{se.alpha}, \code{se.lambda}}{ (vector of) estimated standard
errors of \eqn{\alpha} and \eqn{\lambda} value(s);}

\item{\code{alpha.lambda}}{ numeric, product of \eqn{\alpha} and
\eqn{\lambda} components of the model: this gives the fitted
theoretical minimum success probability of the binary response
in the model;}

\item{\code{se.alpha.lambda}}{ (vector of) estimated standard
errors of \code{alpha.lambda} value(s);}

\item{\code{r.squared}}{ a numeric vector of the same length as
\code{rsquared.method} with pseudo-\eqn{R^2} values returned by
\link[msbreg]{rsquared};}

\item{\code{criterion}}{ the \code{$control$criterion} component of
\code{object};}

\item{\code{alpha.lambda.deficit}}{ logical, does the fit rank suggest that
the \eqn{\alpha} and/or \eqn{\lambda} components are over-parametrized?
Specifically, for a rank deficient information matrix, if the rank
deficit corresponds to the number of parameters in the \eqn{\alpha} and
\eqn{\lambda} components, these parameters are ignored and the rank of
the information for the remaining parameters checked: if the reduced
information is of full rank, then \code{alpha.lambda.deficit = TRUE}
and the variance components for \eqn{\alpha} and/or \eqn{\lambda} are
set to zero. Otherwise, \code{alpha.lambda.deficit = FALSE}.}
}
\description{
Compute summary statistics for a multistage binomial (MSB) model fit.
This is a method for class \code{"msbm"}.
}
\details{
This function mimics the similar \code{summary}
method for "\link[stats]{glm}" class objects.
}
\examples{
data("test1data")

MSBfit = msbreg (y/Total ~ x1 + offset(off1) | x2,
                 alpha.formula = ~ 0,
                 lambda.formula = ~ 0,
                 weights = Total,
                 data = test1data)
MSBfit # Print coefficient estimates (descriptive)

summary (MSBfit) # Print a more detailed summary of the fit (inference)

}
\seealso{
See \link[msbreg]{print.msbm} for printing a quick summary.
Also see \link[msbreg]{msbreg} for model fitting.
}
