% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2data.R
\docType{data}
\name{test2data}
\alias{test2data}
\title{Test Datasets}
\format{
An object of class \code{"data.frame"}.
}
\source{
Simulated
}
\usage{
data(test2data)
}
\description{
Simulated data for testing/demonstrating functions
in the package \code{msbreg}.
}
\details{
The dataset \code{test2data} has 1000 rows and 7 columns.
The outcome \code{y} is a binomial response,
simulated as a Multistage Binomial (MSB) response
using other 6 variables.

The success probability for the \eqn{i}th row is given by:

\eqn{\mu_i = \lambda \times \mu_{i1} \times \mu_{i2} \times \mu_{i3}}

where \eqn{\lambda} is the theoretical maximum success probability,
\eqn{\mu_{ij}} is the success probability at stage \eqn{j} (\eqn{j = 1,2,3}):

\describe{
\item{Stage \eqn{j = 1}}{ has an offset \code{off1} and one numerical predictor \code{x1};}

\item{Stage \eqn{j = 2}}{ has a categorical predictor \code{f2} (3 levels) and one numerical predictor \code{x2};}

\item{Stage \eqn{j = 3}}{ has a offset \code{off3}.}
}

The vector of individual success probabilities can be extracted
as \code{attr(test2data$y, "mu")} and the corresponding vector of
model parameters as \code{attr(test2data$y, "theta")}.
The column \code{test2data$Total} contains the number of trials per subject.

See section \strong{Examples} for MSB model building.
}
\examples{
data(test2data)
head(test2data)

#* MSB model frame
mframe <- msbm.frame (formula = y/Total ~ x1 + offset(off1) | x2 + f2 | offset(off3),
                      alpha.formula = ~ 0,  # minimum probability set to zero
                      lambda.formula = ~ 1, # maximum probability is constant, allowed to be < 1
                      weights = Total,
                      data = test2data, frames = TRUE)

mframe

#* Parameter vector used to simulate the response
parvec <- attr(test2data$y,"theta")
parvec

#* Fisher Information Matrix
fisher.matrix(mframe, coefficients = parvec, link = "logit")

}
\author{
Chenangnon Tovissode \email{ctovissode@uidaho.edu}
}
\keyword{datasets}
