% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcov_msbm.R
\name{vcov.msbm}
\alias{vcov.msbm}
\alias{varcov}
\alias{vcov.summary.msbm}
\title{Variance-Covariance Matrix for \code{'msbm'} Objects}
\usage{
\method{vcov}{msbm}(
  object,
  coefficients = object$coefficients,
  fisher.matrix = TRUE,
  adjust = FALSE,
  sigma = NULL,
  sandwich = FALSE,
  complete = TRUE,
  ...
)

\method{vcov}{summary.msbm}(object, complete = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{'msbm'} (a fitted MSB model as
returned by \link[msbreg]{msbreg}) or class \code{'summary.msbm'} (as
returned by \link[msbreg]{summary.msbm}).}

\item{coefficients}{a numeric vector of all MSB model parameters.
When supplied, \code{coefficients} must have finite value,
and the same length as \code{object$coefficients}.}

\item{fisher.matrix}{logical, should the Fisher (expected) information
matrix be used to obtain the covariance matrix? Defaults to \code{TRUE}.
The alternative is the observed information matrix.}

\item{adjust}{logical, should the covariance matrix be adjusted for
any penalty added to the log-likelihood function during model fitting?
Only relevant when \code{fisher.matrix = FALSE} (Fisher information matrix
is independent of any penalty not involving the response).
Defaults to \code{FALSE}.

When \code{sandwich = TRUE}, \code{adjust = TRUE} also means a finite sample
adjustment made by multiplying the result by \eqn{n/(n−k)} where \eqn{n} is
the number of observations and \eqn{k} the number of estimated parameters.}

\item{sigma}{a numeric scalar (positive real), optional
dispersion parameter, defaults to \code{sigma = 1}.}

\item{sandwich}{logical, should the sandwich estimator be used?
Defaults to \code{FALSE}.}

\item{complete}{logical indicating if the full variance-covariance
matrix should be returned even when some coefficients are aliased (undefined)
and \link[stats]{coef}\code{(.)} contains \code{NA}s correspondingly.
Using \code{complete = TRUE} ensures that \code{vcov(.)} is compatible
with \code{coef(.)}. Setting \code{complete = FALSE} can be useful
to drop rows and columns corresponding to aliased coefficients.}

\item{...}{further arguments passed to or from other methods.
None is currently used.}
}
\value{
A matrix of the estimated variances-covariances for the parameter
in \code{coefficients}.
This has row and column names corresponding to the parameter names
given by the \link[stats]{coef} method.

The returned matrix has the attribute \code{alpha.lambda.deficit}, a
logical indicating whether the information matrix is rank deficient
and the rank deficit is due to the \eqn{\alpha} and/or \eqn{\lambda}
component(s) of the MSB model.
If \code{alpha.lambda.deficit = TRUE}, the rows and columns corresponding
to the \eqn{\alpha} and/or \eqn{\lambda} component(s) of the model fit
in the returned matrix are all set to zero.
}
\description{
Estimate the variance-covariance matrix of a multistage
binomial (MSB) model fit.
}
\details{
These functions mimic and extend the \link[stats]{vcov} method for
"\link[stats]{glm}" class objects to handle \code{'msbm'} objects.
The \code{vcov} method returns a variance-covariance matrix
estimate, by default, the inverse of the information matrix.

For an object of class \code{'msbm'}, the function allows to substitute
the vector of estimated parameters by a user-supplied vector
(\code{coefficients}). It also allows the use of an observed instead
of expected information, and adjusts, if desired, for any  penalty
added to the log-likelihood function during model fitting.

When \code{sandwich = TRUE}, the sandwich covariance matrix is
constructed by multiplying \strong{bread} and \strong{meat} matrices
\insertCite{zeileis2006object}{msbreg}.
The \emph{bread} matrix is the inverse information matrix times the
number of observations, and the \emph{meat} matrix is the average cross
product of the empirical estimating functions (matrix of scores, that
is, first order  derivatives of the log-likelihood contributions
with respect to model parameters, evaluated at \code{coefficients}).
The implementation follows \insertCite{zeileis2020various;textual}{msbreg}.
See \link[msbreg]{bread.meat} for extracting these elementary matrices.
}
\examples{
data("test1data")

MSBfit = msbreg (y/Total ~ x1 + offset(off1) | x2,
                 weights = Total,
                 data = test1data)
summary(MSBfit)

sqrt (diag (vcov (MSBfit)))

round(vcov (MSBfit), 5)

round(vcov (MSBfit, sandwich = TRUE), 5)

}
\references{
\insertAllCited{}
}
\seealso{
See \link[msbreg]{bread.meat} for \code{"msbm"} class methods for
computing empirical estimating functions and their covariance matrix.
}
