% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractscores.R, R/score.test.msbm.R
\name{extractscores}
\alias{extractscores}
\alias{score.test.msbm}
\title{Score Test for Fitted Model Objects}
\usage{
extractscores(object, which = c(1, 3, 5, 6), ...)

\method{score.test}{msbm}(
  object,
  ...,
  scope,
  steps = 1000,
  type = c("expected", "observed"),
  penalty = TRUE,
  identity.link.lambda = FALSE,
  side = NULL,
  trace = object$control$trace,
  ncores = NULL
)
}
\arguments{
\item{object}{an object of a class with a \code{score.test} method.
Typically, a fit, for instance \code{"lm"}, \code{"glm"} and \code{"msbm"}
class objects.}

\item{which}{integer vector, which score types statistic of a
\code{score.test} object should be extracted?
The default selects score statistics based on the Fisher information matrix.
Do not change it unless you are sure about what you are doing.}

\item{...}{further arguments passed to or from other methods.
This may include one or more alternative models of the same class as \code{object};
and arguments such as \code{which} indicating which parameters to test,
\code{type} for using the expected (Fisher) or observed information matrix,
\code{trace} for verbose, and \code{ncores} for the number of cores for
parallel computing.}

\item{scope}{defines the range of alternative models to be examined.
This is used only when \code{...} does not include any model which nests
\code{object}.
This should be either a single formula, or a list containing components
\code{upper} and \code{lower}, both formulae.
See the details for how to specify the formulae and how they are used.}

\item{steps}{the maximum number of null models to be considered.
The default is 1000 (essentially as many as required).
It is typically used to stop the process early when \code{...} is empty
and \code{scope} is a big model.}

\item{type}{character, type of score statistic to compute. The default
(\code{type = "expected"}) uses the Fisher information matrix.
The alternative (\code{type = "observed"}) uses the observed information
matrix.}

\item{penalty}{logical, should the test account for the penalty when a
penalized fitting criterion was used for object?}

\item{identity.link.lambda}{should the identity link function be considered
for the asymptote parameter \eqn{\lambda} in a MSB model?
The default (\code{identity.link.lambda = FALSE}) uses a \code{log} link.

Note that the link used by fitter function is generally a binomial link
function which represents the \eqn{\lambda = 1} as \eqn{\infty} and
results into an identically zero score component for \eqn{\lambda} and
thereby a singular Fisher information matrix.
A different parametrization is thus required/better for \eqn{\lambda} when
performing any of the statistical trinity (score, likelihood ratio, and
Wald tests).}

\item{side}{\code{NULL} \code{NA}, \code{-1}, or \code{1} indicating if
one-sided derivatives should be used when numerical derivative are required
(only relevant when \code{type = "observed"} and \code{penalty = TRUE}).}

\item{trace}{integer indicating if details should be printed during
computations.}

\item{ncores}{integer, number of threads for parallel processing.}
}
\value{
An object of class ‘score.test’ that inherits from the class ‘htest’,
a list with components:

The \code{print} method for \code{"score.test"} class objects calls
\code{scoretable}.

The routine \code{extractscores} returns a vector for a \code{score.test}
input \code{object} and a matrix when \code{object} is a \code{score.list}.
}
\description{
Performs score tests comparing a fitted model to alternatives.
}
\details{
This is a generic function to perform the Rao score test or modified versions
on a model fit object.

The method for an \code{object} of class "\link{lm}" or "\link{glm}" proceeds
as follows.
If a unique model is provided, the default method tests for

When

The function \code{scoretable} generates a summary of one or many
\code{score.test} results.
}
\seealso{
See \link[msbreg]{profile.msbm} for computing log-likelihood profiles
for MSB model fits.
}
